% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_func_mHMM_pois.R
\name{burn_fw}
\alias{burn_fw}
\title{Burn-in the MC chain of each state forward probabilities}
\usage{
burn_fw(x, burn_in = NULL)
}
\description{
Burn-in the MC chain of each state forward probabilities
}
\keyword{internal}
