% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_prob_cpp.R
\name{cont_tv_mult_fw_r_to_cpp}
\alias{cont_tv_mult_fw_r_to_cpp}
\title{Calculates the forward probabilities, used for sampling the state sequence
specific for when data includes a time varying covariate}
\usage{
cont_tv_mult_fw_r_to_cpp(x, m, emiss, n_dep, tgamma, delta = NULL, data_distr)
}
\description{
Calculates the forward probabilities, used for sampling the state sequence
specific for when data includes a time varying covariate
}
