% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_func_mHMM_pois.R
\name{gather_fw}
\alias{gather_fw}
\title{Gather forward probabilities from a nested list into a matrix}
\usage{
gather_fw(x, target)
}
\description{
Gather forward probabilities from a nested list into a matrix
}
\keyword{internal}
