% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_func_mHMM_pois.R
\name{get_aic_pois}
\alias{get_aic_pois}
\title{Obtain the AIC for a fitted multilevel HMM with a Poisson conditional
distribution}
\usage{
get_aic_pois(x, burn_in = NULL)
}
\arguments{
\item{x}{An object of class \code{mHMM_pois}, generated
by the function \code{\link{mHMM_pois}} or \code{\link{mHMM_pois_rm}},
respectively.}

\item{burn_in}{An integer which specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified (\code{burn_in = NULL}), the burn in period specified when
creating the \code{mHMM_pois} or \code{mHMM_pois_rm} object will be used.}
}
\value{
\code{get_aic_pois} returns the AIC for the fitted model averaged
across subjects, based on the number of parameters used in the model and the
loglikelihood. When a repeated measures design has been used, the AIC is
calculated based on the loglikelihood summed over repeated measures of a
same subject, and the AIC averaged across subjects.
}
\description{
\code{get_aic_pois} obtains the Akaike information criterion (AIC) for a
fitted multilevel hidden Markov model with a Poisson state-dependent
conditional distribution, averaged across subjects.
}
