% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_func_mHMM_pois.R
\name{get_map_fw}
\alias{get_map_fw}
\title{Extract the maxima a posteriori (MAPs) for the state-specific forward
probabilities from a fitted multilevel HMM with a Poisson conditional
distribution}
\usage{
get_map_fw(x, burn_in = NULL, target = "median")
}
\arguments{
\item{x}{An object of class \code{mHMM_pois}, generated
by the function \code{\link{mHMM_pois}} or \code{\link{mHMM_pois_rm}},
respectively.}

\item{burn_in}{An integer which specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified (\code{burn_in = NULL}), the burn in period specified when
creating the \code{mHMM_pois} or \code{mHMM_pois_rm} object will be used.}

\item{target}{A character specifying the statistic to be extracted. It
accepts \code{"mean"}, \code{"median"}, or \code{"se"} as inputs. If
no entry is sepcified, it defaults to \code{"median"}.}
}
\value{
\code{get_map_fw} depending on the \code{target} chosen, the
\code{mean}, \code{median} or \code{sd} of each state-specific forward
probability for the full sequence length of all the lowest level units in
a fitted multilevel hidden Markov model with a Poisson state-dependent
conditional distribution is returned.
}
\description{
\code{get_map_fw} extracts the \code{mean}, \code{median} and \code{sd} of
each state-specific forward probability for the full sequence length of all
the lowest level units in a fitted multilevel hidden Markov model with a
Poisson state-dependent conditional distribution. When
\code{\link{mHMM_pois}} has been used to train a model the lowest level
corresponds to subjects, whereas for \code{\link{mHMM_pois_rm}},
repeated measurements.
}
