% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMM_cont.R
\name{mHMM_cont}
\alias{mHMM_cont}
\title{Multilevel hidden  Markov model using Bayesian estimation for continuous
observations}
\usage{
mHMM_cont(
  s_data,
  gen,
  xx = NULL,
  start_val,
  emiss_hyp_prior,
  mcmc,
  return_path = FALSE,
  print_iter,
  show_progress = TRUE,
  gamma_hyp_prior = NULL,
  gamma_sampler = NULL
)
}
\arguments{
\item{s_data}{A matrix containing the observations to be modeled, where the
rows represent the observations over time. In \code{s_data}, the first
column indicates subject id number. Hence, the id number is repeated over
rows equal to the number of observations for that subject. The subsequent
columns contain the dependent variable(s). Note that the dependent
variables are assumed to be continuous (i.e., normally distributed
depending on the hidden states). The total number of rows are equal to the
sum over the number of observations of each subject, and the number of
columns are equal to the number of dependent variables (\code{n_dep}) + 1.
The number of observations can vary over subjects.}

\item{gen}{List containing the following elements denoting the general model
properties:
\itemize{\item{\code{m}: numeric vector with length 1 denoting the number
of hidden states}
\item{\code{n_dep}: numeric vector with length 1 denoting the
number of dependent variables}}}

\item{xx}{An optional list of (level 2) covariates to predict the transition
  matrix and/or the emission probabilities. Level 2 covariate(s) means that
  there is one observation per subject of each covariate. The first element
  in the list \code{xx} is used to predict the transition matrix. Subsequent
  elements in the list are used to predict the emission distribution of (each
  of) the dependent variable(s). Each element in the list is a matrix, with
  the number of rows equal to the number of subjects. The first column of
  each matrix represents the intercept, that is, a column only consisting of
  ones. Subsequent columns correspond to covariates used to predict the
  transition matrix / emission distribution. See \emph{Details} for more
  information on the use of covariates.

  If \code{xx} is omitted completely, \code{xx} defaults to \code{NULL},
  resembling no covariates. Specific elements in the list can also be left
  empty (i.e., set to \code{NULL}) to signify that either the transition
  probability matrix or a specific emission distribution is not predicted by
  covariates.}

\item{start_val}{List containing the start values for the transition
probability matrix gamma and the emission distribution(s). The first
element of the list contains a \code{m} by \code{m} matrix with the start
values for gamma. The subsequent elements are matrices with \code{m} rows
and 2 columns; the first column denoting the mean of state \emph{i} (row
\emph{i}) and the second column denoting the standard deviation of state
\emph{i} (row \emph{i}) of the Normal distribution. Note that
\code{start_val} should not contain nested lists (i.e., lists within
lists).}

\item{emiss_hyp_prior}{An object of the class \code{mHMM_prior_emiss}
containing user specified parameter values for the hyper-prior distribution
on the continuous emission distribution, generated by the function
\code{\link{prior_emiss_cont}}.}

\item{mcmc}{List of Markov chain Monte Carlo (MCMC) arguments, containing the
following elements:
\itemize{\item{\code{J}: numeric vector with length 1 denoting the number
of iterations of the MCMC algorithm}
\item{\code{burn_in}: numeric vector with length 1 denoting the
burn-in period for the MCMC algorithm.}}}

\item{return_path}{A logical scalar. Should the sampled state sequence
obtained at each iteration and for each subject be returned by the function
(\code{sample_path = TRUE}) or not (\code{sample_path = FALSE}). Note that
the sampled state sequence is quite a large object, hence the default
setting is \code{sample_path = FALSE}. Can be used for local decoding
purposes.}

\item{print_iter}{The argument print_iter is depricated; please use
show_progress instead to show the progress of the algorithm.}

\item{show_progress}{A logical scaler. Should the function show a text
progress bar in the \code{R} console to represent the progress of the
algorithm (\code{show_progress = TRUE}) or not (\code{show_progress =
FALSE}). Defaults to \code{show_progress = TRUE}.}

\item{gamma_hyp_prior}{An optional object of class \code{mHMM_prior_gamma}
containing user specified parameter values for the hyper-prior distribution
on the transition probability matrix gamma, generated by the function
\code{\link{prior_gamma}}.}

\item{gamma_sampler}{An optional object of the class \code{mHMM_pdRW_gamma}
containing user specified settings for the proposal distribution of the
random walk (RW) Metropolis sampler on the subject level transition
probability matrix parameters, generated by the function
\code{\link{pd_RW_gamma}}.}
}
\value{
\code{mHMM_cont} returns an object of class \code{mHMM_cont}, which has
  \code{print} and \code{summary} methods to see the results.
  The object contains the following components:
  \describe{
  \item{\code{PD_subj}}{A list containing one matrix per subject with the
  subject level parameter estimates and the log likelihood over the
  iterations of the hybrid Metropolis within Gibbs sampler. The iterations of
  the sampler are contained in the rows, and the columns contain the subject
  level estimates of subsequently the emission means, the (fixed over subjects)
  emission standard deviations, the transition probabilities and the log likelihood.}
  \item{\code{gamma_prob_bar}}{A matrix containing the group level parameter
  estimates of the transition probabilities over the iterations of the hybrid
  Metropolis within Gibbs sampler. The iterations of the sampler are
  contained in the rows, and the columns contain the group level parameter
  estimates. If covariates were included in the analysis, the group level
  probabilities represent the predicted probability given that the covariate
  is at the average value for continuous covariates, or given that the
  covariate equals zero for dichotomous covariates.}
  \item{\code{gamma_int_bar}}{A matrix containing the group level intercepts
  of the multinomial logistic regression modeling the transition
  probabilities over the iterations of the hybrid Metropolis within Gibbs
  sampler. The iterations of the sampler are contained in the rows, and the
  columns contain the group level intercepts.}
  \item{\code{gamma_cov_bar}}{A matrix containing the group level regression
  coefficients of the multinomial logistic regression predicting the
  transition probabilities over the iterations of the hybrid Metropolis within
  Gibbs sampler. The iterations of the sampler are contained in the rows, and
  the columns contain the group level regression coefficients.}
  \item{\code{gamma_int_subj}}{A list containing one matrix per subject
  denoting the subject level intercepts of the multinomial logistic
  regression modeling the transition probabilities over the iterations of the
  hybrid Metropolis within Gibbs sampler. The iterations of the sampler are
  contained in the rows, and the columns contain the subject level
  intercepts.}
  \item{\code{gamma_naccept}}{A matrix containing the number of accepted
  draws at the subject level RW Metropolis step for each set of parameters of
  the transition probabilities. The subjects are contained in the rows, and
  the columns contain the sets of parameters.}
  \item{\code{emiss_mu_bar}}{A list containing one matrix per dependent
  variable, denoting the group level means of the Normal emission
  distribution of each dependent variable over the iterations of the Gibbs
  sampler. The iterations of the sampler are contained in the rows of the
  matrix, and the columns contain the group level emission means. If
  covariates were included in the analysis, the group level means represent
  the predicted mean given that the covariate is at the average value for
  continuous covariates, or given that the covariate equals zero for
  dichotomous covariates.}
  \item{\code{emiss_varmu_bar}}{A list containing one matrix per dependent
  variable, denoting the variance between the subject level means of the
  Normal emission distributions. over the iterations of the Gibbs sampler. The
  iterations of the sampler are contained in the rows of the matrix, and the
  columns contain the group level variance in the mean.}
  \item{\code{emiss_sd_bar}}{A list containing one matrix per dependent
  variable, denoting the (fixed over subjects) standard deviation of the
  Normal emission distributions over the iterations of the Gibbs sampler. The
  iterations of the sampler are contained in the rows of the matrix, and the
  columns contain the group level emission variances.}
  \item{\code{emiss_cov_bar}}{A list containing one matrix per dependent
  variable, denoting the group level regression coefficients predicting the
  emission means within each of the dependent variables over the iterations
  of the Gibbs sampler. The iterations of the sampler are contained in the
  rows  of the matrix, and the columns contain the group level regression
  coefficients.}
  \item{\code{label_switch}}{A matrix of \code{m} rows and \code{n_dep}
  columns containing the percentage of times the group mean of the emission
  distribution of state \code{i} was sampled to be a smaller value compared to
  the group mean of of the emission distribution of state \code{i-1}. If the
  state dependent means of the emission distributions were given in a ranked
  order (low to high) to both the start values and hyper-priors, a high
  percentage in \code{label_switch} indicates that label switching possibly
  poses a problem in the analysis, and further diagnostics (e.g.,
  traceplots and posterior distributions) should be inspected.}
  \item{\code{input}}{Overview of used input specifications: the number of
  states \code{m}, the number of used dependent variables \code{n_dep}, the
  number of iterations \code{J} and the specified burn in period
  \code{burn_in} of the hybrid Metropolis within Gibbs sampler, the number of
  subjects \code{n_subj}, the observation length for each subject
  \code{n_vary}, and the column names of the dependent variables
  \code{dep_labels}.}
  \item{\code{sample_path}}{A list containing one matrix per subject with the
  sampled hidden state sequence over the hybrid Metropolis within Gibbs
  sampler. The time points of the dataset are contained in the rows, and the
  sampled paths over the iterations are contained in the columns. Only
  returned if \code{return_path = TRUE}. }
}
}
\description{
\code{mHMM_cont} fits a multilevel (also known as mixed or random effects)
hidden Markov model (HMM) to intense longitudinal data with continuous
observations (i.e., normally distributed) of multiple subjects using Bayesian
estimation, and creates an object of class mHMM_cont. By using a multilevel
framework, we allow for heterogeneity in the model parameters between
subjects, while estimating one overall HMM. The function includes the
possibility to add covariates at level 2 (i.e., at the subject level) and
have varying observation lengths over subjects. For a short description of
the package see \link{mHMMbayes}. See \code{vignette("tutorial-mhmm")} for an
introduction to multilevel hidden Markov models and the package, and see
\code{vignette("estimation-mhmm")} for an overview of the used estimation
algorithms.
}
\details{
Covariates specified in \code{xx} can either be dichotomous or continuous
variables. Dichotomous variables have to be coded as 0/1 variables.
Categorical or factor variables can as yet not be used as predictor
covariates. The user can however break up the categorical variable in
multiple dummy variables (i.e., dichotomous variables), which can be used
simultaneously in the analysis. Continuous predictors are automatically
centered. That is, the mean value of the covariate is subtracted from all
values of the covariate such that the new mean equals zero. This is done such
that the presented probabilities in the output (i.e., for the population
transition probability matrix and population emission probabilities)
correspond to the predicted probabilities at the average value of the
covariate(s).
}
\examples{
###### Example on simulated data
# simulating multivariate continuous data
n_t     <- 100
n       <- 10
m       <- 3
n_dep   <- 2

gamma   <- matrix(c(0.8, 0.1, 0.1,
                    0.2, 0.7, 0.1,
                    0.2, 0.2, 0.6), ncol = m, byrow = TRUE)

emiss_distr <- list(matrix(c( 5, 1,
                              10, 1,
                              15, 1), nrow = m, byrow = TRUE),
                    matrix(c(0.5, 0.2,
                             1.0, 0.5,
                             2.0, 0.3), nrow = m, byrow = TRUE))

data_cont <- sim_mHMM(n_t = n_t, n = n, gen = list(m = m, n_dep = n_dep), data_distr = 'continuous',
                  gamma = gamma, emiss_distr = emiss_distr, var_gamma = .1, var_emiss = c(.5, 0.01))

# Specify hyper-prior for the continuous emission distribution
manual_prior_emiss <- prior_emiss_cont(
                        gen = list(m = m, n_dep = n_dep),
                        emiss_mu0 = list(matrix(c(3.0, 7.0, 17.0), nrow = 1),
                                         matrix(c(0.7, 0.8, 1.8), nrow = 1)),
                        emiss_K0 = list(1, 1),
                        emiss_V =  list(rep(2, m), rep(1, m)),
                        emiss_nu = list(1, 1),
                        emiss_a0 = list(rep(1, m), rep(1, m)),
                        emiss_b0 = list(rep(1, m), rep(1, m)))

# Run the model on the simulated data:
# Note that for reasons of running time, J is set at a ridiculous low value.
# One would typically use a number of iterations J of at least 1000,
# and a burn_in of 200.
out_3st_cont_sim <- mHMM_cont(s_data = data_cont$obs,
                    gen = list(m = m, n_dep = n_dep),
                    start_val = c(list(gamma), emiss_distr),
                    emiss_hyp_prior = manual_prior_emiss,
                    mcmc = list(J = 11, burn_in = 5))


}
\references{
\insertRef{rabiner1989}{mHMMbayes}

\insertRef{scott2002}{mHMMbayes}

\insertRef{altman2007}{mHMMbayes}

\insertRef{rossi2012}{mHMMbayes}

\insertRef{zucchini2017}{mHMMbayes}
}
\seealso{
\code{\link{sim_mHMM}} for simulating multilevel hidden Markov data,
  \code{\link{vit_mHMM}} for obtaining the most likely hidden state sequence
  for each subject using the Viterbi algorithm, \code{\link{obtain_gamma}}
  and \code{\link{obtain_emiss}} for obtaining the transition or emission
  distribution probabilities of a fitted model at the group or subject level,
  and \code{\link{plot.mHMM}} for plotting the posterior densities of a
  fitted model.
}
