% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMM_pois.R
\name{mHMM_pois}
\alias{mHMM_pois}
\title{Multilevel hidden Markov model using Bayesian estimation for count
observations}
\usage{
mHMM_pois(
  s_data,
  gen,
  xx = NULL,
  start_val,
  emiss_hyp_prior,
  mcmc,
  return_path = FALSE,
  return_fw_prob = FALSE,
  print_iter,
  show_progress = TRUE,
  gamma_hyp_prior = NULL,
  gamma_sampler = NULL,
  alpha_scale = 1
)
}
\arguments{
\item{s_data}{A matrix containing the observations to be modelled, where the
rows represent the observations over time. In \code{s_data}, the first
column indicates subject id number. Hence, the id number is repeated over
rows equal to the number of observations for that subject. The subsequent
columns contain the dependent variable(s). Note that the dependent
variables are assumed to be continuous (i.e., normally distributed
depending on the hidden states). The total number of rows are equal to the
sum over the number of observations of each subject, and the number of
columns are equal to the number of dependent variables (\code{n_dep}) + 1.
The number of observations can vary over subjects.}

\item{gen}{List containing the following elements denoting the general model
properties:
\itemize{\item{\code{m}: numeric vector with length 1 denoting the number
of hidden states}
\item{\code{n_dep}: numeric vector with length 1 denoting the
number of dependent variables}}}

\item{xx}{An optional list of (level 2) covariates to predict the transition
  matrix. Level 2 covariate(s) means that
  there is one observation per subject of each covariate. The first element
  in the list \code{xx} is used to predict the transition matrix. Subsequent
  elements in the list are used to predict the emission distribution of (each
  of) the dependent variable(s). Each element in the list is a matrix, with
  the number of rows equal to the number of subjects. The first column of
  each matrix represents the intercept, that is, a column only consisting of
  ones. Subsequent columns correspond to covariates used to predict the
  transition matrix. See \emph{Details} for more
  information on the use of covariates. Support for prediction of emission
  probabilities may be added on future versions of the package.

  If \code{xx} is omitted completely, \code{xx} defaults to \code{NULL},
  resembling no covariates. Specific elements in the list can also be left
  empty (i.e., set to \code{NULL}) to signify that either the transition
  probability matrix or a specific emission distribution is not predicted by
  covariates.}

\item{start_val}{List containing the start values for the transition
probability matrix gamma and the emission distribution(s). The first
element of the list contains a \code{m} by \code{m} matrix with the start
values for gamma. The subsequent elements are matrices with \code{m} rows
and 1 column denoting the Poisson parameters lambda of state \emph{i} (row
\emph{i}). Note that \code{start_val}
should not contain nested lists (i.e., lists within lists).}

\item{emiss_hyp_prior}{A list containing user specified parameters
 of the hyper-prior distribution on the Poisson emission
 distributions (and regression coefficients given that covariates are used)
 for each of the states. The hyper-prior connected to the means of the
 Poisson emission distribution(s) is a hierarchical Poisson-Gamma
 distribution. The hyper-priors on the scale (\code{emiss_alpha_bar}) and
 rate (\code{emiss_beta_bar}) of the Gamma prior
 are defined by Gamma hyper-priors with hyper-parameters
 \code{emiss_alpha_a0}, \code{emiss_alpha_b0}, \code{emiss_beta_a0} and
 \code{emiss_beta_b0}, respectively.

 Hence, the list \code{emiss_hyp_prior} contains the following elements:
 \itemize{\item{\code{emiss_alpha_a0}: a list containing \code{n_dep} elements corresponding
 to each of the dependent variables \code{k}, where each element is a vector
 with lenght \code{m} containing the shape values of the Gamma
 hyper-prior on each of the means of the Poisson emission distribuitons
 (note: here the standard Gamma parametrization is used, with parameters shape and rate).}
 \item{\code{emiss_alpha_b0}: a list containing \code{n_dep} elements corresponding
 to each of the dependent variables \code{k}, where each element is a vector
 with lenght \code{m} containing the rate values of the Gamma
 hyper-prior on each of the means of the Poisson emission distribuitons
 (note: here the standard Gamma parametrization is used, with parameters shape and rate).}
 \item{\code{emiss_beta_a0}: a list containing \code{n_dep} elements corresponding
 to each of the dependent variables \code{k}, where each element is a vector
 with lenght \code{m} containing the shape values of the Gamma
 hyper-prior on each of the means of the Poisson emission distribuitons
 (note: here the standard Gamma parametrization is used, with parameters shape and rate).}
 \item{\code{emiss_beta_b0}: a list containing \code{n_dep} elements corresponding
 to each of the dependent variables \code{k}, where each element is a vector
 with lenght \code{m} containing the rate values of the Gamma
 hyper-prior on each of the means of the Poisson emission distribuitons
 (note: here the standard Gamma parametrization is used, with parameters shape and rate).}}


 See \emph{Details} below if covariates are used for changes in the settings
 of the arguments of \code{emiss_hyp_prior}.}

\item{mcmc}{List of Markov chain Monte Carlo (MCMC) arguments, containing the
following elements:
\itemize{\item{\code{J}: numeric vector with length 1 denoting the number
of iterations of the MCMC algorithm}
\item{\code{burn_in}: numeric vector with length 1 denoting the
burn-in period for the MCMC algorithm.}}}

\item{return_path}{A logical scalar. Should the sampled state sequence
obtained at each iteration and for each subject be returned by the function
(\code{sample_path = TRUE}) or not (\code{sample_path = FALSE}). Note that
the sampled state sequence is quite a large object, hence the default
setting is \code{sample_path = FALSE}. Can be used for local decoding
purposes.}

\item{return_fw_prob}{A logical scalar. Should the states forward
probabilities obtained at each time step and each iteration for each
subject be returned by the function (\code{return_fw_prob = TRUE}) or not
(\code{return_fw_prob = FALSE}). Note that the forward probability
sequence is quite a large object, hence the default setting is
\code{return_fw_prob = FALSE}.}

\item{print_iter}{The argument print_iter is depricated; please use
show_progress instead to show the progress of the algorithm.}

\item{show_progress}{A logical scaler. Should the function show a text
progress bar in the \code{R} console to represent the progress of the
algorithm (\code{show_progress = TRUE}) or not (\code{show_progress =
FALSE}). Defaults to \code{show_progress = TRUE}.}

\item{gamma_hyp_prior}{An optional list containing user specified parameters
 of the hyper-prior distribution on the multivariate normal distribution
 of the intercepts (and regression coefficients given that covariates are
 used) of the multinomial regression model of the transition probability
 matrix gamma. The hyper-prior of the mean intercepts is a multivariate
 Normal distribution, the hyper-prior of the covariance matrix between the
 set of (state specific) intercepts is an Inverse Wishart distribution.

 Hence, the list \code{gamma_hyp_prior} contains the following elements:
 \itemize{\item{\code{gamma_mu0}: a list containing m matrices; one matrix
 for each row of the transition probability matrix gamma. Each matrix
 contains the hypothesized mean values of the intercepts. Hence, each matrix
 consists of one row (when not including covariates in the model) and
 \code{m} - 1 columns}
 \item{\code{gamma_K0}: numeric vector with length 1 denoting the number of
 hypothetical prior subjects on which the vector of means \code{gamma_mu0} is
 based}
 \item{\code{gamma_nu}: numeric vector with length 1 denoting the degrees of
 freedom of the Inverse Wishart distribution}
 \item{\code{gamma_V}: matrix of \code{m} - 1 by \code{m} - 1 containing the
 hypothesized variance-covariance matrix between the set of intercepts.}}
 Note that \code{gamma_K0}, \code{gamma_nu} and \code{gamma_V} are assumed
 equal over the states. The mean values of the intercepts (and regression
 coefficients of the covariates) denoted by \code{gamma_mu0} are allowed to
 vary over the states.

 The default values for the hyper-prior on gamma are: all elements of the
 matrices contained in \code{gamma_mu0} set to 0, \code{gamma_K0} set to 1,
 \code{gamma_nu} set to 3 + m - 1, and the diagonal of \code{gamma_V} (i.e.,
 the variance) set to 3 + m - 1 and the off-diagonal elements (i.e., the
 covariance) set to 0.

 See \emph{Details} below if covariates are used for changes in the settings
 of the arguments of \code{gamma_hyp_prior}.}

\item{gamma_sampler}{An optional list containing user specified settings for
 the proposal distribution of the random walk (RW) Metropolis sampler for
 the subject level parameter estimates of the intercepts modeling the
 transition probability matrix. The list \code{gamma_sampler} contains the
 following elements:
\itemize{\item{\code{gamma_int_mle0}: a numeric vector with length \code{m}
- 1 denoting the start values for the maximum likelihood estimates of the
intercepts for the transition probability matrix gamma, based on the pooled
data (data over all subjects)}
\item{\code{gamma_scalar}: a numeric vector with length 1 denoting the scale
factor \code{s}. That is, The scale of the proposal distribution is composed
of a covariance matrix Sigma, which is then tuned by multiplying it by a
scaling factor \code{s}^2}
\item{\code{gamma_w}: a numeric vector with length 1 denoting the weight for
the overall log likelihood (i.e., log likelihood based on the pooled data
over all subjects) in the fractional likelihood.}}
 Default settings are: all elements in \code{gamma_int_mle0} set to 0,
 \code{gamma_scalar} set to 2.93 / sqrt(\code{m} - 1), and \code{gamma_w} set to
 0.1. See the section \emph{Scaling the proposal distribution of the RW
 Metropolis sampler} in \code{vignette("estimation-mhmm")} for details.}

\item{alpha_scale}{A numeric value specifying the scalar used on the
Random-Walk Metropolis within Gibbs used to sample the population-level
shape parameter alpha. This is the tuning parameter that can be used to
ensure an efficient sampling from the posterior distribution of
\code{alpha}. \code{alpha_scale} defaults to 1.}
}
\value{
\code{mHMM_pois} returns an object of class \code{mHMM_pois}, which has
  \code{print} and \code{summary} methods to see the results.
  The object contains the following components:
  \describe{
  \item{\code{PD_subj}}{A list containing one matrix per subject with the
  subject level parameter estimates and the log likelihood over the
  iterations of the hybrid Metropolis within Gibbs sampler. The iterations of
  the sampler are contained in the rows, and the columns contain the subject
  level estimates of subsequently the emission means, the (fixed over subjects)
  emission variances, the transition probabilities and the log likelihood.}
  \item{\code{gamma_prob_bar}}{A matrix containing the group level parameter
  estimates of the transition probabilities over the iterations of the hybrid
  Metropolis within Gibbs sampler. The iterations of the sampler are
  contained in the rows, and the columns contain the group level parameter
  estimates. If covariates were included in the analysis, the group level
  probabilities represent the predicted probability given that the covariate
  is at the average value for continuous covariates, or given that the
  covariate equals zero for dichotomous covariates.}
  \item{\code{gamma_int_bar}}{A matrix containing the group level intercepts
  of the multinomial logistic regression modeling the transition
  probabilities over the iterations of the hybrid Metropolis within Gibbs
  sampler. The iterations of the sampler are contained in the rows, and the
  columns contain the group level intercepts.}
  \item{\code{gamma_cov_bar}}{A matrix containing the group level regression
  coefficients of the multinomial logistic regression predicting the
  transition probabilities over the iterations of the hybrid Metropolis within
  Gibbs sampler. The iterations of the sampler are contained in the rows, and
  the columns contain the group level regression coefficients.}
  \item{\code{gamma_int_subj}}{A list containing one matrix per subject
  denoting the subject level intercepts of the multinomial logistic
  regression modeling the transition probabilities over the iterations of the
  hybrid Metropolis within Gibbs sampler. The iterations of the sampler are
  contained in the rows, and the columns contain the subject level
  intercepts.}
  \item{\code{gamma_naccept}}{A matrix containing the number of accepted
  draws at the subject level RW Metropolis step for each set of parameters of
  the transition probabilities. The subjects are contained in the rows, and
  the columns contain the sets of parameters.}
  \item{\code{emiss_naccept}}{A matrix containing the number of accepted
  draws at the subject level RW Metropolis step for each set of
  \code{alpha_bar} of the emission probabilities. The dependent variables
  are contained in the rows, and the columns contain the hidden states.}
  \item{\code{emiss_alpha_bar}}{A list containing one matrix per dependent
  variable, denoting the group level means of the Normal emission
  distribution of each dependent variable over the iterations of the Gibbs
  sampler. The iterations of the sampler are contained in the rows of the
  matrix, and the columns contain the group level emission means. If
  covariates were included in the analysis, the group level means represent
  the predicted mean given that the covariate is at the average value for
  continuous covariates, or given that the covariate equals zero for
  dichotomous covariates.}
  \item{\code{emiss_beta_bar}}{A list containing one matrix per dependent
  variable, denoting the variance between the subject level means of the
  Normal emision distributions. over the iterations of the Gibbs sampler. The
  iterations of the sampler are contained in the rows of the matrix, and the
  columns contain the group level variance in the mean.}
  \item{\code{input}}{Overview of used input specifications: the number of
  states \code{m}, the number of used dependent variables \code{n_dep}, the
  number of iterations \code{J} and the specified burn in period
  \code{burn_in} of the hybrid Metropolis within Gibbs sampler, the number of
  subjects \code{n_subj}, the observation length for each subject
  \code{n_vary}, and the column names of the dependent variables
  \code{dep_labels}.}
  \item{\code{sample_path}}{A list containing one matrix per subject with the
  sampled hidden state sequence over the hybrid Metropolis within Gibbs
  sampler. The time points of the dataset are contained in the rows, and the
  sampled paths over the iterations are contained in the columns. Only
  returned if \code{return_path = TRUE}. }
  \item{\code{forward_prob}}{A nested list containing two or more matrices
  per subject with the sampled forward probabilities of a given state \code{i}
  over the iterations. The time points of the dataset are contained in the
  rows, and the sampled forward probabilities for a stae \code{i} over the
  iterations are contained in the columns. Only returned if
  \code{return_fw_prob = TRUE}. }
}
}
\description{
\code{mHMM_pois} fits a multilevel (also known as mixed or random effects)
hidden Markov model (HMM) to intense longitudinal data with count
observations (i.e., Poisson distributed) of multiple subjects using Bayesian
estimation, and creates an object of class mHMM_pois. By using a multilevel
framework, we allow for heterogeneity in the model parameters between
subjects, while estimating one overall HMM. The function includes the
possibility to add covariates at level 2 (i.e., at the subject level) and
have varying observation lengths over subjects. For a short description of
the package see \link{mHMMbayes}. See \code{vignette("tutorial-mhmm")} for an
introduction to multilevel hidden Markov models and the package, and see
\code{vignette("estimation-mhmm")} for an overview of the used estimation
algorithms.
}
\details{
Covariates specified in \code{xx} can either be dichotomous or continuous
variables. Dichotomous variables have to be coded as 0/1 variables.
Categorical or factor variables can as yet not be used as predictor
covariates. The user can however break up the categorical variable in
multiple dummy variables (i.e., dichotomous variables), which can be used
simultaneously in the analysis. Continuous predictors are automatically
centered. That is, the mean value of the covariate is subtracted from all
values of the covariate such that the new mean equals zero. This is done such
that the presented probabilities in the output (i.e., for the population
transition probability matrix and population emission probabilities)
correspond to the predicted probabilities at the average value of the
covariate(s).

If covariates are specified and the user wants to set the values for the
parameters of the hyper-prior distributions manually, the specification of
the elements in the arguments of the hyper-prior parameter values of the
normal distribution on the means for the transition probabilities
change as follows: the number of rows in the
matrix \code{gamma_mu0} is equal to 1 + the number of
covariates used to predict the transition probability matrix for
\code{gamma_mu0} (i.e.,
the first row correspond to the hyper-prior mean values of the intercepts,
the subsequent rows correspond to the hyper-prior mean values of the
regression coefficients connected to each of the covariates), and
\code{gamma_K0} is now a matrix with the number of
hypothetical prior subjects on which the vectors of means (i.e., the rows in
\code{gamma_mu0}) are based on the diagonal, and
off-diagonal elements equal to 0. Note that the hyper-prior parameter values
of the inverse Wishart distribution on the covariance matrix remains
unchanged, as the estimates of the regression coefficients for the covariates
are fixed over subjects.
}
\examples{
###### Example on simulated data
# simulating multivariate continuous data
n_t     <- 400
n       <- 30
m       <- 2
n_dep   <- 3

gamma   <- matrix(c(0.95, 0.05,
                    # 0.2, 0.7, 0.1,
                    0.02, 0.98), ncol = m, byrow = TRUE)

emiss_distr <- list(matrix(c(20, 1,
                             # 1, 1,
                             1, 1), nrow = m, byrow = TRUE),
                    matrix(c(15, 1,
                             # 2, 1,
                             5, 1), nrow = m, byrow = TRUE),
                    matrix(c(50, 1,
                             # 3, 1,
                             20, 1), nrow = m, byrow = TRUE))

set.seed(42)
data_pois <- sim_mHMM_pois(n_t = n_t, n = n, m = m, n_dep = n_dep, data_distr = 'poisson',
                           gamma = gamma, emiss_distr = emiss_distr, var_gamma = .1, var_emiss = NULL, return_ind_par = TRUE)

# Specify hyper-prior for the continuous emission distribution
hyp_pr <- list(
  emiss_alpha_a0 = rep(list(rep(1, m)),n_dep),
  emiss_alpha_b0 = rep(list(rep(1, m)),n_dep),
  emiss_beta_a0 = rep(list(rep(1, m)),n_dep),
  emiss_beta_b0 = rep(list(rep(1, m)),n_dep)
)

# Specify starting values
start_gamma <- diag(0.8, m)
start_gamma[lower.tri(start_gamma) | upper.tri(start_gamma)] <- (1 - diag(start_gamma)) / (m - 1)

start_emiss <- vector("list",length = n_dep)
for(i in 1:n_dep) {start_emiss[[i]] <- matrix(c(30,0.1), nrow = m, byrow = TRUE)}

# Run the model on the simulated data:
out <- mHMM_pois(s_data = data_pois$obs,
                 gen = list(m = m, n_dep = n_dep),
                 start_val = c(list(start_gamma), start_emiss),
                 emiss_hyp_prior = hyp_pr,
                 mcmc = list(J = 100, burn_in = 50), alpha_scale = 10, return_path = TRUE, return_fw_prob = TRUE)



}
\references{
\insertRef{rabiner1989}{mHMMbayes}

\insertRef{scott2002}{mHMMbayes}

\insertRef{altman2007}{mHMMbayes}

\insertRef{rossi2012}{mHMMbayes}

\insertRef{zucchini2017}{mHMMbayes}
}
\seealso{
\code{\link{sim_mHMM}} for simulating multilevel hidden Markov data,
  \code{\link{vit_mHMM}} for obtaining the most likely hidden state sequence
  for each subject using the Viterbi algorithm, \code{\link{obtain_gamma}}
  and \code{\link{obtain_emiss}} for obtaining the transition or emission
  distribution probabilities of a fitted model at the group or subject level,
  and \code{\link{plot.mHMM}} for plotting the posterior densities of a
  fitted model.
}
