% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMMbayes.R
\docType{package}
\name{mHMMbayes}
\alias{mHMMbayes}
\alias{mHMMbayes-package}
\title{mHMMbayes: A package for multilevel hidden Markov models using Bayesian estimation.}
\description{
The mHMMbayes package provides two main functions: mHMM_mnl and sim_mHMM
}
\section{mHMM_mnl}{

The function mHMM_mnl is used to analyses (intense longitudinal) data from
multiple subjects using a multilevel hidden Markov model. By using a
multilevel framework, one general 'population' HMM is estimated, while
heterogeneity between subjects is accommodated. The function can handle
covariates at the subject level (unlimited number), uses a hybrid metropolis
within gibs sampler, and performs the forward backward algorithm for all
subjects in a sequential manner. Can handle varying observation length over
subjects.
}

\section{sim_mHMM}{

The function sim_mHMM is used to simulate data for multiple subjects, for
which the data have categorical observations that follow a hidden Markov
model (HMM) with an multilevel structure. The multilevel structure implies
that each subject is allowed to have it's own set of parameters, and that the
parameters at the subject level (level 1) are tied together by a population
distribution at level 2 for each of the corresponding parameters. The shape
of the population distribution for each of the parameters is a normal (i.e.,
Gaussian) distribution. In addition to (natural and/or unexplained)
heterogeneity between subjects, the subjects parameters can also depend on a
(set of) covariate(s).
}

