% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonverbal}
\alias{nonverbal}
\title{Nonverbal communication of patients and therapist}
\format{A matrix with 10 * 900 rows and 5 variables:
\describe{
\item{id}{id variable of patient - therapis couple to distinguish which observation belongs to which couple}
\item{p_verbalizing}{verbalizing behavior of the patient, consisiting of 1 = verbalizing, 2 = back chanelling, 3 = not verbalizing}
\item{p_looking}{looking behavior of the patient, consisting of 1 = looking at therapist, 2 = not looking at therapist}
\item{t_verbalizing}{verbalizing behavior of the therapist, consisiting of 1 = verbalizing, 2 = back chanelling, 3 = not verbalizing}
\item{t_looking}{looking behavior of the therapist, consisting of 1 = looking at patient, 2 = not looking at patient}
}}
\source{
Give source here once published, or something else referring to
  earlier published work on data
}
\usage{
nonverbal
}
\description{
A dataset containing the nonverbal communication of 10 patient-therapist couples,
recorded for 15 minutes at a frequency of 1 observation per second (= 900 observations per couple).
}
\keyword{datasets}
