% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonverbal_cov}
\alias{nonverbal_cov}
\title{Predictors of nonverbal communication}
\format{
A matrix with 10 rows and 3 variables:
\describe{
\item{diagnosis}{Diagnosis of the patient, consisting of 0 = depression, 1 = anxiety}
\item{std_CDI_change}{Change in measure for depression (CDI) before and after therapy, standardized scale}
\item{std_SCA_change}{Change in measure for anxiety (SCARED) before and after therapy, standardized scale}
}
}
\usage{
nonverbal_cov
}
\description{
A dataset containing predictors of nonverbal communication of 10 patient-therapist couples.
}
\keyword{datasets}
