% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_gamma.R
\name{obtain_gamma}
\alias{obtain_gamma}
\title{Obtain the transition probabilities gamma for a fitted multilevel HMM}
\usage{
obtain_gamma(object, level = "group", burn_in = NULL)
}
\arguments{
\item{object}{An object of class \code{mHMM_f}, generated by the function
\code{\link{mHMM_f}}.}

\item{level}{String specifying if the returned transition probability matrix
gamma should be at the group level (\code{level = "group"}), i.e.,
representing the average transition probability matrix over all subjects,
or at the subject level (\code{level = "subject"}).}

\item{burn_in}{An integer which specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified (\code{burn_in = NULL}), the burn in period specified when
creating the \code{mHMM_f} object will be used.}
}
\value{
\code{obtain_gamma} creates an object of the class \code{mHMM_gamma}.
  This object can be directly plotted using the function
  \code{plot.mHMM_gamma()}, or simply \code{plot()}. Depending on the
  specification at the input variable \code{level}, the output is either a
  matrix with the transition probabilities at the group level (if \code{level
  = "group"}), or a list of matrices (with the number of elements equal to
  the number of subjects analyzed, if \code{level = 'subject'}), where each
  matrix in the list represents a subject specific transition probability
  matrix.
}
\description{
\code{obtain_gamma} obtains the transition probability matrix (TPM) for an
  object containing a fitted multilevel hidden Markov model. The TPM can be
  obtained either at the group level, i.e., representing the average transition
  probability matrix over all subjects, or at the subject level, returning the
  transition probability matrices for each subject.
}
\examples{
###### currently no example

}
\seealso{
\code{\link{mHMM_f}} for fitting the multilevel hidden Markov model,
  and \code{\link{plot.mHMM_gamma}} for plotting the obtained transition
  probabilities.
}
