% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mHMM.R
\name{plot.mHMM}
\alias{plot.mHMM}
\title{Plotting the posterior densities for a fitted multilevel HMM}
\usage{
\method{plot}{mHMM}(x, component = "gamma", dep = 1, col, cat_lab,
  dep_lab, lwd1 = 2, lwd2 = 1, lty1 = 1, lty2 = 3, legend_cex,
  burn_in, ...)
}
\arguments{
\item{x}{Object of class \code{mHMM}, generated by the function
\code{\link{mHMM_mnl}}.}

\item{component}{String specifying if the displayed posterior densities
should be for the transition probablity matrix gamma (\code{component =
"gamma"}), or for the emission distribution probabilities (\code{component
= "emiss"}). In case of the latter and the model is based on multiple
dependent variables, the user has to indicate for which dependent variable
the posterior densities have to be plotted, see \code{dep}.}

\item{dep}{Integer specifying for which dependent variable the posterior
desities should be plotted. Only required if one wishes to plot the
emission distriubtion probabilites and the model is based on multiple
dependent variables.}

\item{col}{Vector of colors for the posterior density lines.}

\item{cat_lab}{Vector of strings denoting the labels of the categorical outcome
values. Automatically genereted when not provided. Only required if one
wishes to plot the emission distriubtion probabilites.}

\item{dep_lab}{Strings denoting the label for the dependent variable plotted.
Automatically obtained from the input object x when not specified. Only
required if one wishes to plot the emission distriubtion probabilites.}

\item{lwd1}{Positive number indicating the line width of the posterior
density at the group level.}

\item{lwd2}{Positive number indicating the line width of the posterior
density at the subject level.}

\item{lty1}{Positive number indicating the line type of the posterior
density at the group level.}

\item{lty2}{Positive number indicating the line type of the posterior
density at the subject level.}

\item{legend_cex}{A numerical value giving the amount by which plotting text
and symbols in the legend should be magnified relative to the default.}

\item{burn_in}{An integer wich specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified, the burn in period specified at creating the \code{mHMM}
object with the function \code{\link{mHMM_mnl}} will be used.}

\item{...}{Arguments to be passed to methods (see \code{\link[graphics]{par}})}
}
\value{
\code{plot.mHMM} returns a plot of the transition probability
  matrix. Depending on whether the input object represents the transiton
  probabilities at the group level or the subject specific transition
  probability matrices, the returned plot represents either the group
  transition probality matrix, or the transition probability matrix for a
  given subject, specified by \code{subject_nr}.
}
\description{
\code{plot.mHMM} plots the posterior densities for a fitted multilevel hidden
markov model for the group and subject level parameters simultaneously. The
plotted posterior densities are either for the transition probability matrix
gamma, or for the emission distribution probabilities.
}
\seealso{
\code{\link{mHMM_mnl}} for fitting the multilevel hidden Markov
  model, creating the object \code{mHMM}.
}
