% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mHMM_cont.R
\name{plot.mHMM_cont}
\alias{plot.mHMM_cont}
\title{Plotting the posterior densities for a fitted multilevel HMM}
\usage{
\method{plot}{mHMM_cont}(
  x,
  component = "gamma",
  dep = 1,
  col,
  dep_lab,
  lwd1 = 2,
  lwd2 = 1,
  lty1 = 1,
  lty2 = 3,
  legend_cex,
  burn_in,
  ...
)
}
\arguments{
\item{x}{Object of class \code{mHMM_cont}, generated by the function
\code{\link{mHMM}}.}

\item{component}{String specifying if the displayed posterior densities
should be for the transition probability matrix gamma (\code{component =
"gamma"}), or for the emission distribution probabilities (\code{component
= "emiss"}). In case of the latter and the model is based on multiple
dependent variables, the user has to indicate for which dependent variable
the posterior densities have to be plotted, see \code{dep}.}

\item{dep}{Integer specifying for which dependent variable the posterior
densities should be plotted. Only required if one wishes to plot the
emission distribution probabilities and the model is based on multiple
dependent variables.}

\item{col}{Vector of colors for the posterior density lines. If one is
plotting the posterior densities for gamma, the vector has length \code{m}
(i.e., number of hidden states). If one is plotting the posterior densities
for the emission probabilities, the vector has length \code{q_emiss[k]}
(i.e., the number of outcome categories for the dependent variable
\code{k}).}

\item{dep_lab}{Optional vector of strings when plotting the posterior
densities of the emission probabilities with length \code{n_dep}, denoting
the labels of the dependent variables. Automatically obtained from the
input object \code{x} when not specified.}

\item{lwd1}{Positive number indicating the line width of the posterior
density at the group level.}

\item{lwd2}{Positive number indicating the line width of the posterior
density at the subject level.}

\item{lty1}{Positive number indicating the line type of the posterior
density at the group level.}

\item{lty2}{Positive number indicating the line type of the posterior
density at the subject level.}

\item{legend_cex}{A numerical value giving the amount by which plotting text
and symbols in the legend should be magnified relative to the default.}

\item{burn_in}{An integer which specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified, the burn in period specified at creating the \code{mHMM}
object with the function \code{\link{mHMM}} will be used.}

\item{...}{Arguments to be passed to methods (see \code{\link[graphics]{par}})}
}
\value{
\code{plot.mHMM_cont} returns a plot of the posterior densities. Depending
  on whether (\code{component = "gamma"}) or (\code{component = "emiss"}),
  the plotted posterior densities are either for the transition probability
  matrix gamma or for the emission distribution probabilities, respectively.
}
\description{
\code{plot.mHMM_cont} plots the posterior densities for a fitted multilevel hidden
Markov model for the group and subject level parameters simultaneously. The
plotted posterior densities are either for the transition probability matrix
gamma, or for the emission distribution probabilities.
}
\examples{
###### First run the function mHMM on the nonverbal data
\donttest{
# specifying general model properties:
}

}
\seealso{
\code{\link{mHMM_cont}} for fitting the multilevel hidden Markov
  model, creating the object \code{mHMM_cont}.
}
