% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mHMM_emiss.R
\name{plot.mHMM_emiss}
\alias{plot.mHMM_emiss}
\title{Plotting the emission distribution for a fitted multilevel HMM}
\usage{
\method{plot}{mHMM_emiss}(
  x,
  subj_nr = NULL,
  by_state = TRUE,
  cat_lab,
  dep_lab,
  col,
  legend_cex = 1,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mHMM_emiss}, generated by the function
\code{\link{obtain_emiss}}.}

\item{subj_nr}{An integer specifying for which specific subject the
transition probability matrix should be plotted. Only required if the input
object represents the subject specific transition probability matrices.}

\item{by_state}{An optional logic variable. It indicates how user would like
to display the plots. It is set to \code{TRUE} by default and it
display plots state wise. If dependent variable wise presentation
is preferred \code{by_state} should be set to \code{FALSE}
(\code{by_state=FALSE})}

\item{cat_lab}{optional list of n_dep(number of dependent variables) vectors
of strings when plotting emission probability distribution variables in input
object x. The first vector represents labels of the dependent variable's
categories which is listed first in the input object x. To provide only some
labels of dependent variables, user need to set the rest list elements to
\code{NA}. Labels are automatically obtained from the input object x when not
specified.}

\item{dep_lab}{Optional vector of strings when plotting emission probability
distribution, denoting the labels of dependent variable. Length of the vector
need to be of the number of dependent variables in input object x. The first
position of the vector is a label of the first object in the input object x.
To provide only chosen labels user need to set the rest to \code{NA}.
When a position of the vector is set to \code{NA}, the label is generated from the
input object x. Labels are automatically obtained from the input object x
when not specified.}

\item{col}{An optional vector of the size of the number of categories of the
dependent variable that has them the most,specifying the used colors in the
\code{\link[graphics]{barplot}}.}

\item{legend_cex}{A numerical value giving the amount by which
to the default.}

\item{...}{Arguments to be passed to methods
(see \code{\link[graphics]{barplot}})}
}
\value{
\code{plot.mHMM_gamma} returns a plot of the emission probability
  matrix. Depending on whether the input object represents list of emission
  probability matrices at the group level or the subject specific list of
  emission probability matrices, the returned plot represents either
  the group emission probability distribution for each state, or the emission
  probability distribution for each state of given subject,
  specified by \code{subject_nr}.
}
\description{
\code{plot.mHMM_emiss} plots the emission probability distribution for
a fitted multilevel hidden Markov model.
The plotted emission probability distributions either represents
the probabilities at the group level, i.e., representing the average emission
probability distribution over all subjects, or at the subject level.
In case of the latter, the user has to specify for which subject the emission
probability matrix should be plotted.
}
\examples{
\donttest{
m <- 3 #number of states
n_dep <- 4 #number of dependent variables used to infer the hidden states
q_emiss <- c(3, 2, 3, 2) #number of categorical outcomes of
                          each of dependent variables
start_TM <- diag(.8, m) #transition matrix
start_TM
start_TM[lower.tri(start_TM) | upper.tri(start_TM)] <- .2
start_EM <- list(matrix(c(0.05, 0.90, 0.05,
                          0.90, 0.05, 0.05,
                         0.60, 0.30, 0.10), byrow = TRUE,
                       nrow = m, ncol = q_emiss[1]), # vocalizing patient
                matrix(c(0.1, 0.9,
                         0.1, 0.9,
                         0.5, 0.5), byrow = TRUE, nrow = m,
                       ncol = q_emiss[2]), # looking patient
                matrix(c(0.90, 0.05, 0.05,
                         0.05, 0.90, 0.05,
                         0.60, 0.30, 0.10), byrow = TRUE,
                       nrow = m, ncol = q_emiss[3]), # vocalizing therapist
                matrix(c(0.1, 0.9,
                         0.1, 0.9,
                         0.5, 0.5), byrow = TRUE, nrow = m,
                       ncol = q_emiss[4])) # looking therapist

out_3st <- mHMM(s_data = nonverbal,
                gen = list(m = m, n_dep = n_dep, q_emiss = q_emiss),
                start_val = c(list(start_TM), start_EM),
                mcmc = list(J = 10, burn_in = 5))

# creating object of obtain_emiss class
group_emiss3<-obtain_emiss(out_3st,level = "group")

subject_emiss3<-obtain_emiss(out_3st,level = "subject")

#plotting
plot(x=subject_emiss3,subj_nr = 2)

plot(x=group_emiss3)

plot(x=group_emiss3,by_state=FALSE)


}
\dontrun{
plot(x=subject_emiss3,subj_nr = 2,by_state = F,col = c("red","orange",
     "green"))
plot(x=subject_emiss3,subj_nr = 7,by_state = T,col = c("red","orange","green"),
    dep_lab=c("Patient Looking",NA,"Therapist Vocalizing",
               "Therapist Looking"),cat_lab=list())

plot(x=subject_emiss3,subj_nr = 7,by_state = T,col = c("red","orange","green"),
     dep_lab=c("Patient Looking",NA,"Therapist Vocalizing",
     "Therapist Looking"),cat_lab=list(c("low","medium","high"),NA,NA,c("1","3")))
}

}
\seealso{
\code{\link{mHMM}} for fitting the multilevel hidden Markov
  model, creating the object \code{mHMM}, and \code{\link{obtain_emiss}} to
  obtain the transition probabilities gamma for a fitted multilevel HMM,
  creating the object \code{mHMM_gamma}.
}
