% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Plotting the predicted transition/emission probabilities given certain values of covariates}
\usage{
plot_pred(
  object,
  component = "gamma",
  cov = 1,
  dep = 1,
  col,
  cat_lab,
  dep_lab,
  cov_lab,
  ...
)
}
\arguments{
\item{object}{Object of class \code{mHMM}, generated by the function
\code{\link{mHMM}}.}

\item{component}{String specifying if the plot should be made for the
transition probability matrix gamma (\code{component = "gamma"}) or
for the emission distribution probabilities (\code{component = "emiss"}).}

\item{cov}{Integer that specifies which covariate to use for plotting when
there are multiple covariates available. The default value is 1, which
indicates the first covariate in the list of covariates.}

\item{dep}{Integer specifying for which dependent variable the predicted
emission probabilities should be plotted.}

\item{col}{Vector of colors for the plots. If one is plotting the predicted
transition probabilities, the vector has length \code{m} (i.e., number of
hidden states). If one is plotting the predicted emission probabilities,
the vector has length \code{q_emiss[k]} (i.e., the number of outcome
categories for the dependent variable \code{k}).
If not specified, colors will be assigned automatically.}

\item{cat_lab}{Optional vector of strings when plotting the predicted
emission probabilities, denoting the labels of the categorical outcome values.
Automatically generated when not provided.}

\item{dep_lab}{Optional string when plotting the predicted the emission
probabilities with length 1, denoting the label for the dependent variable
plotted. Automatically obtained from the input object when not specified.}

\item{cov_lab}{Optional string denoting the label for the covariate variable
plotted. Automatically generated when not provided.}

\item{...}{Other parameters passed down to \code{plot_pred()}.}
}
\description{
Plotting the predicted transition/emission probabilities given certain values of covariates
}
