% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_prob.R
\name{pred_prob}
\alias{pred_prob}
\title{Compute predicted transition/emission probabilities given certain values of covariates}
\usage{
pred_prob(object, component = "gamma", dep = 1, print.df = TRUE, ...)
}
\arguments{
\item{object}{Object of class \code{mHMM}, generated by the function
\code{\link{mHMM}}.}

\item{component}{String specifying if the prediction should be made
for the transition probability matrix gamma (\code{component = "gamma"}),
or for the emission distribution probabilities (\code{component = "emiss"}).
In case of the latter and the model is based on multiple dependent variables,
the user has to indicate a specific dependent variable, see \code{dep}.}

\item{dep}{Integer specifying for which dependent variable the predicted e
mission probabilities are computed.}

\item{print.df}{A logical argument specifying whether the output should be
returned as a dataframe (`TRUE`) or lists (`FALSE`). Default is `TRUE`.}

\item{...}{Other parameters passed down to \code{pred_probs()}.}
}
\value{
Either a dataframe or lists of dataframes containing predicted
   transition/emission probabilities.
}
\description{
Compute predicted transition/emission probabilities given certain values of covariates
}
