% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_gamma.R
\name{prior_gamma}
\alias{prior_gamma}
\title{Specifying informative hyper-prior on the transition probability matrix gamma of the multilevel hidden Markov model}
\usage{
prior_gamma(
  m,
  gamma_mu0,
  gamma_K0 = NULL,
  gamma_nu = NULL,
  gamma_V = NULL,
  n_xx_gamma = NULL
)
}
\arguments{
\item{m}{Numeric vector with length 1 denoting the number of hidden states.}

\item{gamma_mu0}{A list containing m matrices; one matrix for each row of the
transition probability matrix gamma. Each matrix contains the hypothesized
hyper-prior mean values of the intercepts of the Multinomial logit
model on the transition probabilities gamma. Hence, each matrix
consists of one row (when not including covariates in the model) and
\code{m} - 1 columns. If covariates are used, the number of rows in each
matrix in the list is equal to 1 + n_xx_gamma (i.e., the first row corresponds to
the hyper-prior mean values of the intercepts, the subsequent rows
correspond to the hyper-prior mean values of the regression coefficients
connected to each of the covariates).}

\item{gamma_K0}{Optional numeric vector with length 1 (when no covariates are
used) denoting the number of hypothetical prior subjects on which the set
of hyper-prior mean intercepts specified in \code{gamma_mu0} are based.
When covariates are used: Numeric vector with length 1 + n_xx_gamma denoting the
number of hypothetical prior subjects on which the set of intercepts (first
value) and set of regression coefficients (subsequent values) are based.}

\item{gamma_nu}{Optional numeric vector with length 1 denoting the degrees of freedom
of the hyper-prior Inverse Wishart distribution on the covariance of
the Multinomial logit intercepts.}

\item{gamma_V}{Optional matrix of \code{m} - 1 by \code{m} - 1 containing the
variance-covariance matrix of the hyper-prior Inverse Wishart distribution
on the covariance of the Multinomial logit intercepts.}

\item{n_xx_gamma}{Optional numeric vector with length 1 denoting the number
of (level 2) covariates (excluding the intercept) used to predict the
transition probability matrix gamma. When omitted, the model assumes no
covariates are used to predict gamma.}
}
\value{
\code{prior_gamma} returns an object of class \code{mHMM_prior_gamma},
  containing informative hyper-prior values for the transition probability
  matrix gamma of the multilevel hidden Markov model. The object is
  specifically created and formatted for use by the function \code{mHMM},
  and thoroughly checked for correct input dimensions.
  The object contains the following components:
  \describe{
  \item{\code{m}}{Numeric vector denoting the number of hidden states, used
  for checking equivalent general model properties specified under
  \code{prior_gamma} and \code{mHMM}.}
  \item{\code{gamma_mu0}}{A list containing the hypothesized hyper-prior mean
  values of the intercepts of the Multinomial logit model on the transition
  probability matrix gamma.}
  \item{\code{gamma_K0}}{A numeric vector denoting the number of hypothetical
  prior subjects on which the set of hyper-prior mean intercepts specified in
  \code{gamma_mu0} are based.}
  \item{\code{gamma_nu}}{A numeric vector denoting the degrees of freedom
  of the hyper-prior Inverse Wishart distribution on the covariance of the
  Multinomial logit intercepts.}
  \item{\code{gamma_V}}{A matrix containing the variance-covariance of the
  hyper-prior Inverse Wishart distribution on the covariance of the
  Multinomial logit intercepts.}
  \item{\code{n_xx_gamma}}{A numeric vector denoting the number of (level 2)
  covariates used to predict the transition probability matrix gamma. When no
  covariates are used, \code{n_xx_gamma} equals \code{NULL}.}
  }
}
\description{
\code{prior_gamma} provides a framework to manually specify an informative
hyper-prior on the transition probability matrix gamma, and creates an object
of class \code{mHMM_prior_gamma} used by the function \code{mHMM_f}. Note that the
hyper-prior distribution on the transition probabilities are on the
intercepts (and, if subject level covariates are used, regression
coefficients) of the Multinomial logit model used to accommodate the
multilevel framework of the data, instead of on the probabilities directly.
The set of hyper-prior distributions consists of a multivariate Normal
hyper-prior distribution on the vector of means (i.e., intercepts and
regression coefficients), and an Inverse Wishart hyper-prior distribution on
the covariance matrix.
}
\details{
Estimation of the mHMM proceeds within a Bayesian context, hence a
hyper-prior distribution has to be defined for the group level parameters.
Default, non-informative priors are used unless specified otherwise by the
user. Each row of the transition probability matrix has its own set of
Multinomial logit intercepts, which are assumed to follow a multivariate
normal distribution. Hence, the hyper-prior distributions for the intercepts
consists of a multivariate Normal hyper-prior distribution on the vector of
means, and an Inverse Wishart hyper-prior distribution on the covariance
matrix. Note that only the number of states \code{m} and values of the
hypothesized hyper-prior mean values of the Multinomial logit intercepts have
to be specified by the user, default values are available for all other
hyper-prior distribution parameters.

Given that the hyper-priors are specified on the intercepts of the Multinomial
logit model intercepts instead of on the probabilities of the transition
probability matrix gamma directly, specifying a hyper-prior can seem rather
daunting. However, see the function \code{\link{prob_to_int}} and
\code{\link{int_to_prob}} for translating probabilities to a set of
Multinomial logit intercepts and vice versa.

Note that \code{gamma_K0}, \code{gamma_nu} and \code{gamma_V} are assumed
equal over the states. When the hyper-prior values for \code{gamma_K0},
\code{gamma_nu} and \code{gamma_V} are not manually specified, the default
values are as follows. \code{gamma_K0} set to 1, \code{gamma_nu} set to 3 + m
- 1, and the diagonal of \code{gamma_V} (i.e., the variance) set to 3 + m - 1
and the off-diagonal elements (i.e., the covariance) set to 0. In addition,
when no manual values for the hyper-prior on gamma are specified at all (that
is, the function \code{prior_gamma} is not used), all elements of the
matrices contained in \code{gamma_mu0} are set to 0 in the function
\code{mHMM_f}.

Note that in case covariates are specified, the hyper-prior parameter values
of the inverse Wishart distribution on the covariance matrix remain
unchanged, as the estimates of the regression coefficients for the covariates
are fixed over subjects.
}
\seealso{
\code{\link{prob_to_int}} for transforming a set of probabilities to a set
  of Multinomial logit regression intercepts, and \code{\link{mHMM_f}} for
  fitting a multilevel hidden Markov model.
}
