% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mHMM_v.R
\name{sim_mHMM_v}
\alias{sim_mHMM_v}
\title{Simulate data using a multilevel hidden Markov model}
\usage{
sim_mHMM_v(
  n_t,
  n,
  data_distr = "categorical",
  gen,
  gamma,
  emiss_distr,
  start_state = NULL,
  xx_vec = NULL,
  beta = NULL,
  var_gamma = 0.1,
  var_emiss = NULL,
  return_ind_par = FALSE,
  m,
  n_dep
)
}
\arguments{
\item{n_t}{Numeric vector with length 1 denoting the length of the observed
sequence to be simulated for each subject. To only simulate subject
specific transition probability matrices gamma and emission distributions
(and no data), set \code{t} to 0.}

\item{n}{Numeric vector with length 1 denoting the number of subjects for
which data is simulated.}

\item{data_distr}{A character vector of length 1 denoting the distribution
adopted for the data given the hidden states. In this version of the
package, it can only take the default value 'continuous', standing for the
continuous observations following a normal distribution.}

\item{gen}{List containing the following elements denoting the general model
properties:
\itemize{
\item{\code{m}: numeric vector with length 1 denoting the number
of hidden states}
\item{\code{n_dep}: numeric vector with length 1 denoting the
number of dependent variables}}}

\item{gamma}{A matrix with \code{m} rows and \code{m} columns containing the
average population transition probability matrix used for simulating the
data. That is, the probability to switch from hidden state \emph{i} (row
\emph{i}) to hidden state \emph{j} (column  \emph{j}).}

\item{emiss_distr}{A list with \code{n_dep} elements containing the average
population emission distribution(s) of the observations given the hidden
states for each of the dependent variables.Each element is a matrix with
\code{m} rows and 2 columns; the first column denoting the mean of state
\emph{i} (row \emph{i}) and the second column denoting the standard
deviation of state \emph{i} (row \emph{i}) of the Normal distribution.}

\item{start_state}{Optional numeric vector with length 1 denoting in which
state the simulated state sequence should start. If left unspecified, the
simulated state for time point 1 is sampled from the initial state
distribution (which is derived from the transition probability matrix
gamma).}

\item{xx_vec}{List of 1 + \code{n_dep} vectors containing the covariate(s) to
predict the transition probability matrix \code{gamma} and/or (specific)
emission distribution(s) \code{emiss_distr} using the regression parameters
specified in \code{beta} (see below). The first element in the list
\code{xx_vec} is used to predict the transition matrix. Subsequent elements
in the list are used to predict the emission distribution of (each of) the
dependent variable(s). This means that the covariate used to predict
\code{gamma} and \code{emiss_distr} can either be the same covariate,
different covariates, or a covariate for certain elements and none for the
other. At this point, it is only possible to use one covariate for both
\code{gamma} and \code{emiss_distr}. For all elements in the list,
the number of observations in the vectors should be  equal to the number of
subjects to be simulated \code{n}. If \code{xx_vec} is omitted completely,
\code{xx_vec} defaults to NULL, resembling no covariates at all. Specific
elements in the list can also be left empty (i.e., set to \code{NULL}) to
signify that either the transition probability matrix or (one of) the
emission distribution(s) is not predicted by covariates.}

\item{beta}{List of 1 + \code{n_dep} matrices containing the regression
  parameters to predict \code{gamma} and/or \code{emiss_distr} in combination
  with \code{xx_vec} using (Multinomial logistic) regression. The first
  matrix is used to predict the transition probability matrix \code{gamma}.
  The subsequent matrices are used to predict the emission distribution(s)
  \code{emiss_distr} of the dependent variable(s). For \code{gamma}, one
  regression parameter is specified for each element in \code{gamma}, with
  the following exception. The first element in each row of \code{gamma} is
  used as reference category in the Multinomial logistic regression. As such,
  no regression parameters can be specified for these parameters. Hence, the
  first element in the list \code{beta} to predict \code{gamma} consist of a
  matrix with the number of rows equal to \code{m} and the number of columns
  equal to \code{m} - 1.  See \emph{details} for more information. For
  continuous and count emission distributions, the subsequent elements in the
  list \code{beta} consist of a matrix with the number of rows equal to
  \code{m} and 1 column.

  Note that if \code{beta} is specified, \code{xx_vec} has to be specified as
  well. If \code{beta} is omitted completely, \code{beta} defaults to NULL,
  resembling no prediction of \code{gamma} and \code{emiss_distr} using
  covariates. One of the elements in the list can also be left empty
  (i.e., set to \code{NULL}) to signify that either the transition
  probability matrix or a specific emission distribution is not predicted by
  covariates.}

\item{var_gamma}{Either a numeric vector with length 1 or a matrix of
(\code{m} by \code{m} - 1) elements denoting the amount of variance
between subjects in the transition probability matrix. Note that the
value(s) correspond to the variance of the parameters of the Multinomial
distribution (i.e., the intercepts of the regression equation of the
Multinomial distribution used to sample the transition probability
matrix), see details below. Also note that if only one variance value is
provided, it will be adopted for the complete transition probability
matrix, hence the variance is assumed fixed across all components. The
default equals 0.1, which corresponds to little variation between
subjects. If one wants to simulate data from exactly the same HMM for all
subjects, var_gamma should be set to 0. Note that if data for only 1
subject is simulated (i.e., n = 1), \code{var_gamma} is set to 0.}

\item{var_emiss}{Either a numeric vector with length \code{n_dep} or a list
of \code{n_dep} matrices denoting the amount of variance between subjects
in the emission distribution(s). When opting for a list of matrices: each
element is a matrix with \code{m} rows and 1 column; the single column
denoting the variance between the subjects' means of state \emph{i} (row
\emph{i}) of the normal distribution used as prior for the Normal
emissions. This value corresponds to the variance in the mean of the
emission distribution(s) across subjects. Note that if only one variance
value provided for each emission distribution, the variance is assumed
fixed across states within  an emission distribution. The default equals
0.1. Note that if data for only 1 subject is simulated (i.e., n = 1),
\code{var_emiss} is set to a vector of 0's.}

\item{return_ind_par}{A logical scalar. Should the subject specific
transition probability matrix \code{gamma} and emission probability matrix
\code{emiss_distr} be returned by the function (\code{return_ind_par =
TRUE}) or not (\code{return_ind_par = FALSE}). The default equals
\code{return_ind_par = FALSE}.}

\item{m}{The argument \code{m} is deprecated; please specify using the input
parameter \code{gen}.}

\item{n_dep}{The argument \code{n_dep} is deprecated; please specify using
the input parameter \code{gen}.}
}
\value{
The following components are returned by the function \code{sim_mHMM}:
\describe{
  \item{\code{states}}{A matrix containing the simulated hidden state
  sequences, with one row per hidden state per subject. The first column
  indicates subject id number. The second column contains the simulated
  hidden state sequence, consecutively for all subjects. Hence, the id number
  is repeated over the rows (with the number of repeats equal to the length
  of the simulated hidden state sequence \code{T} for each subject).}
  \item{\code{obs}}{A matrix containing the simulated observed outputs, with
  one row per simulated observation per subject. The first column indicates
  subject id number. The second column contains the simulated observation
  sequence, consecutively for all subjects. Hence, the id number is repeated
  over rows (with the number of repeats equal to the length of the simulated
  observation sequence \code{T} for each subject).}
  \item{\code{gamma}}{A list containing \code{n} elements with the simulated
  subject specific transition probability matrices \code{gamma}. Only
  returned if \code{return_ind_par} is set to \code{TRUE}.}
  \item{\code{emiss_distr}}{A list containing \code{n} elements with the
  simulated subject specific emission probability matrices
  \code{emiss_distr}. Only returned if \code{return_ind_par} is set to
  \code{TRUE}.}
}
}
\description{
\code{sim_mHMM_v} simulates data for multiple subjects, for which the data have
continuous (i.e., normally distributed)
observations that follow a hidden Markov model (HMM) with a
multilevel structure. The multilevel structure implies that each subject is
allowed to have its own set of parameters, and that the parameters at the
subject level (level 1) are tied together by a population distribution at
level 2 for each of the corresponding parameters. The shape of the population
distribution for each of the parameters is a normal distribution. In addition
to (natural and/or unexplained) heterogeneity between subjects, the subjects
parameters can also depend on a covariate.
}
\details{
In simulating the data, having a multilevel structure means that the
parameters for each subject are sampled from the population level
distribution of the corresponding parameter. The user specifies the
population distribution for each parameter: the average population transition
probability matrix and its variance, and the average population emission
distribution and its variance. For now, the variance of the mean population
parameters is assumed fixed for all components of the transition probability
matrix and for all components of the emission distribution.

One can simulate multivariate data. That is, the hidden states depend on more
than 1 observed variable simultaneously. The distributions of multiple
dependent variables for multivariate data are assumed to be independent.

Note that the subject specific initial state distributions (i.e., the
probability of each of the states at the first time point) needed to simulate
the data are obtained from the stationary distributions of the subject
specific transition probability matrices gamma.

\code{beta}: As the first element in each row of \code{gamma} is used as
reference category in the Multinomial logistic regression, the first matrix
in the list \code{beta} used to predict transition probability matrix
\code{gamma} has a number of rows equal to \code{m} and the number of columns
equal to \code{m} - 1. The first element in the first row corresponds to the
probability of switching from state one to state two. The second element in
the first row corresponds to the probability of switching from state one to
state three, and so on. The last element in the first row corresponds to the
probability of switching from state one to the last state.
}
\examples{

## Example on multivariate continuous data
# simulating multivariate continuous data
n_t     <- 100
n       <- 10
m       <- 3
n_dep   <- 2

gamma   <- matrix(c(0.8, 0.1, 0.1,
                    0.2, 0.7, 0.1,
                    0.2, 0.2, 0.6), ncol = m, byrow = TRUE)

emiss_distr <- list(matrix(c( 50, 10,
                              100, 10,
                              150, 10), nrow = m, byrow = TRUE),
                    matrix(c(5, 2,
                             10, 5,
                             20, 3), nrow = m, byrow = TRUE))

data_cont <- sim_mHMM_v(n_t = n_t, n = n, data_distr = 'continuous',
                      gen = list(m = m, n_dep = n_dep),
                      gamma = gamma, emiss_distr = emiss_distr,
                      var_gamma = .5, var_emiss = c(5^2, 0.2^2))

head(data_cont$states)
head(data_cont$obs)


}
\seealso{
\code{\link{mHMM_f}} for analyzing multilevel hidden Markov data.
}
