% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Plotting the traceplot(s) corresponding to one or more Markov chains}
\usage{
traceplot(L, component = "gamma", dep = 1, col, cat_lab, dep_lab, burn_in, ...)
}
\arguments{
\item{L}{a list of objects of class \code{mHMM} generated by the
\code{\link{mHMM}} function.}

\item{component}{A string is used to specify whether the traceplot should be
created for the parameter estimates of the transition probabilities
(\code{component = "gamma"}) or the emission probabilities
(\code{component = "emiss"}). When the component is `"emiss"` and
the model is based on multiple dependent variables, the user must indicate
a specific dependent variable to be plotted using dep, see \code{dep}.}

\item{dep}{Integer specifying for which dependent variable the traceplot
 should be plotted. Only required if one wishes to plot the
emission distribution probabilities and the model is based on multiple
dependent variables.}

\item{col}{Optional vector of colors for the traceplot lines. The vector
should have a length equal to the number of \code{mHMM} objects in the
list \code{L}. If not specified, colors will be automatically assigned.}

\item{cat_lab}{Optional vector of strings when plotting for the emission
probabilities, denoting the labels of the categorical outcome values.
Automatically generated when not provided.}

\item{dep_lab}{Optional string when plotting for the emission probabilities,
denoting the label for the dependent variable plotted.
Automatically obtained from the input object \code{x} when not specified.}

\item{burn_in}{An integer which specifies the number of iterations to discard.
When left unspecified, the burn in period specified at creating the
\code{mHMM} object with the function \code{\link{mHMM}} will be used.}

\item{...}{Other parameters passed down to \code{traceplot()}.}
}
\description{
\code{traceplot.mHMM} creates the traceplots for the group-level parameter
   estimates corresponding to a fitted multilevel hidden Markov model.
}
