% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vit_mHMM_cont.R
\name{vit_mHMM_cont}
\alias{vit_mHMM_cont}
\title{Obtain hidden state sequence for each subject using the Viterbi
algorithm}
\usage{
vit_mHMM_cont(object, s_data, burn_in = NULL)
}
\arguments{
\item{object}{An object of class \code{mHMM_cont}, generated by the
function \code{\link{mHMM_cont}}, respectively.}

\item{s_data}{A matrix containing the observations to be modelled, where the
rows represent the observations over time. In \code{s_data}, the first
column indicates subject id number. Hence, the id number is repeated over
rows equal to the number of observations for that subject. The subsequent
columns contain the dependent variable(s). Note that the dependent
variables have to be numeric, i.e., they cannot be a (set of) factor
variable(s). The total number of rows are equal to the sum over the number
of observations of each subject, and the number of columns are equal to the
number of dependent variables (\code{n_dep}) + 1. The number of
observations can vary over subjects.}

\item{burn_in}{The number of iterations to be discarded from the MCMC
algorithm when inferring the transition probability matrix gamma and the
emission distribution of (each of) the dependent variable(s) for each
subject from \code{s_data}. If omitted, defaults to \code{NULL} and
\code{burn_in} specified at the function \code{mHMM} will be used.}
}
\value{
The function \code{vit_mHMM_cont} returns a matrix containing the most
  likely state at each point in time. Each column represents a subject, and
  each row represents a point in time. If sequence lengths differ over
  subjects, states for none existing time points for subjects are filled with
  \code{NA}.
}
\description{
\code{vit_mHMM_cont} obtains the most likely state sequence (for each subject)
from an object of class \code{mHMM_cont} (generated by the
function \code{mHMM_cont()} for sequences obtained from multiple subjects or animals,
respectively), using (an extended version of) the Viterbi algorithm. This is
also known as global decoding.
}
\details{
Note that local decoding is also possible, by inferring the most frequent
state at each point in time for each subject from the sampled state path at
each iteration of the MCMC algorithm. This information is contained in the
output object \code{return_path} of the function \code{mHMM_cont()}.
}
