% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_mHMM1}
\alias{forecast_mHMM1}
\title{Forecasts (forward predictions) using a multilevel hidden Markov model}
\usage{
forecast_mHMM1(s_data, object, forecast_steps = 1, burn_in = NULL)
}
\description{
Ideas:
(1) use MAPs similarly as done with the viterbi algorithm:
 +++ fast
 --- no measure of uncertainty on forcasts
(2) obtain forecasts over iterations (after burn-in):
 +++ uncertainty on forecasts
 --- slow
}
