% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{HMM}
\alias{HMM}
\title{Onelevel hidden  Markov model using Bayesian estimation}
\usage{
HMM(s_data, gen, start_val, mcmc, return_path = FALSE,
  show_progress = TRUE, gamma_prior = NULL, emiss_prior = NULL)
}
\arguments{
\item{s_data}{A matrix containing the observations to be modelled, where the
rows represent the observations over time. In \code{s_data}, the first
column indicates subject id number. Hence, the id number is repeated over
rows equal to the number of observations for that subject. The subsequent
columns contain the dependent variable(s). Note that the dependent
variables have to be numeric, i.e., they cannot be a (set of) factor
variable(s). The total number of rows are equal to the sum over the number
of observations of each subject, and the number of columns are equal to the
number of dependent variables (\code{n_dep}) + 1. The number of
observations can vary over subjects.}

\item{xx}{sfsfd}
}
\value{
\code{HMM} returns an object of class \code{HMM}, which has
  \code{print} and \code{summary} methods to see the results.
  The object contains the following components:
  \describe{
  \item{\code{xx}}{A .. containing ...}
}
}
\description{
\code{HMM} fits a onelevel (i.e., fixed parameter) hidden Markov model (HMM)
to intense longitudinal data with categorical observations using Bayesian
estimation, and creates an object of class HMM. For a short description of
the package see \link{mHMMbayes}. See \code{vignette("tutorial-mhmm")} for an
introduction to multilevel hidden Markov models and the package, and see
\code{vignette("estimation-mhmm")} for an overview of the used estimation
algorithms.
}
\examples{

}
\seealso{

}
