% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMM_f.R
\name{mHMM_f}
\alias{mHMM_f}
\title{Multilevel hidden  Markov model using Bayesian estimation}
\usage{
mHMM_f(
  s_data,
  data_distr = "continuous",
  gen,
  xx = NULL,
  start_val,
  mcmc,
  return_path = FALSE,
  show_progress = TRUE,
  gamma_hyp_prior = NULL,
  emiss_hyp_prior = NULL,
  gamma_sampler = NULL
)
}
\arguments{
\item{s_data}{A matrix containing the observations to be modeled, where the
rows represent the observations over time. In \code{s_data}, the first
column indicates subject id number. Hence, the id number is repeated over
rows equal to the number of observations for that subject. The subsequent
columns contain the dependent variable(s). The total number of rows are
equal to the sum over the number of observations of each subject, and the
number of columns are equal to the number of dependent variables
(\code{n_dep}) + 1. The number of observations can vary over subjects.}

\item{data_distr}{String vector with length 1 describing the
observation type of the data. Currently supported are \code{'categorical'}
, \code{'continuous'}, and \code{'count'}. Note that for multivariate
data, all dependent variables are assumed to be of the same observation
type. The default equals to \code{data_distr = 'categorical'}.}

\item{gen}{List containing the following elements denoting the general model
properties:
\itemize{
\item{\code{m}: numeric vector with length 1 denoting the number
of hidden states}
\item{\code{n_dep}: numeric vector with length 1 denoting the
number of dependent variables}}}

\item{xx}{An optional list of (level 2) covariates to predict the transition
  matrix. Level 2 covariate(s) means that there is one observation per
  subject of each covariate. The first element in the list \code{xx} is used
  to predict the transition matrix. Each element in the list is a matrix,
  with the number of rows equal to the number of subjects. The first column
  of each matrix represents the intercept, that is, a column only consisting
  of ones. Subsequent columns correspond to covariates used to predict the
  transition matrix / emission distribution. See \emph{Details} for more
  information on the use of covariates. Note that since the emission means
  are fixed over subjects in this version, no covariates can be used to
  predict variation in the emission means over subjects. Hence, this list
  here always contains only 1 element.

  If \code{xx} is omitted completely, \code{xx} defaults to \code{NULL},
  resembling no covariates.}

\item{start_val}{List containing the start values for the transition
probability matrix gamma and the emission distribution(s). The first
element of the list contains a \code{m} by \code{m} matrix with the start
values for gamma. The subsequent \code{n_dep} elements each contain a
matrix with the start values for the emission distribution(s): a \code{m}
by 2 matrix denoting the mean (first column) and standard deviation (second
column) of the Normal emission distribution within each state (rows). Note
that \code{start_val} should not contain nested lists (i.e., lists within
lists).}

\item{mcmc}{List of Markov chain Monte Carlo (MCMC) arguments, containing the
following elements:
\itemize{\item{\code{J}: numeric vector with length 1 denoting the number
of iterations of the MCMC algorithm}
\item{\code{burn_in}: numeric vector with length 1 denoting the
burn-in period for the MCMC algorithm.}}}

\item{return_path}{A logical scalar. Should the sampled state sequence
obtained at each iteration and for each subject be returned by the function
(\code{sample_path = TRUE}) or not (\code{sample_path = FALSE}). Note that
the sampled state sequence is quite a large object, hence the default
setting is \code{sample_path = FALSE}. Can be used for local decoding
purposes.}

\item{show_progress}{A logical scaler. Should the function show a text
progress bar in the \code{R} console to represent the progress of the
algorithm (\code{show_progress = TRUE}) or not (\code{show_progress =
FALSE}). Defaults to \code{show_progress = TRUE}.}

\item{gamma_hyp_prior}{An optional object of class \code{mHMM_prior_gamma}
containing user specified parameter values for the hyper-prior distribution
on the transition probability matrix gamma, generated by the function
\code{\link{prior_gamma}}.}

\item{emiss_hyp_prior}{An object of the class \code{mHMM_prior_emiss}
containing user specified parameter values for the hyper-prior distribution
on the (fixed over subjects) Normal emission distribution(s), generated by
the function \code{\link{prior_emiss_cont_f}}.}

\item{gamma_sampler}{An optional object of the class \code{mHMM_pdRW_gamma}
containing user specified settings for the proposal distribution of the
random walk (RW) Metropolis sampler on the subject level transition
probability matrix parameters, generated by the function
\code{\link{pd_RW_gamma}}.}
}
\value{
\code{mHMM_f} returns an object of class \code{mHMM_f}, which has
  \code{print} and \code{summary} methods to see the results.
  The object contains the following components:
  \describe{
  \item{\code{PD_subj}}{A list containing one list per subject with the
  elements \code{trans_prob} and \code{log_likl}, providing the subject
  parameter estimates over the iterations of the MCMC sampler.
  \code{trans_prob} relates to the transition probabilities gamma,
  and \code{log_likl} to the log likelihood over the MCMC iterations.
  Iterations are contained in the rows, the parameters in the columns.}
  \item{\code{gamma_prob_bar}}{A matrix containing the group level parameter
  estimates of the transition probabilities over the iterations of the hybrid
  Metropolis within Gibbs sampler. The iterations of the sampler are
  contained in the rows, and the columns contain the group level parameter
  estimates. If covariates were included in the analysis, the group level
  probabilities represent the predicted probability given that the covariate
  is at the average value for continuous covariates, or given that the
  covariate equals zero for dichotomous covariates.}
  \item{\code{gamma_int_bar}}{A matrix containing the group level intercepts
  of the Multinomial logistic regression modeling the transition
  probabilities over the iterations of the hybrid Metropolis within Gibbs
  sampler. The iterations of the sampler are contained in the rows, and the
  columns contain the group level intercepts.}
  \item{\code{gamma_cov_bar}}{A matrix containing the group level regression
  coefficients of the Multinomial logistic regression predicting the
  transition probabilities over the iterations of the hybrid Metropolis within
  Gibbs sampler. The iterations of the sampler are contained in the rows, and
  the columns contain the group level regression coefficients.}
  \item{\code{gamma_V_int_bar}}{A matrix containing the variance
  components for the subject-level intercepts (between subject variances)
  of the multinomial logistic regression modeling the transition
  probabilities over the iterations of the hybrid Metropolis within Gibbs
  sampler. The iterations of the sampler are contained in the rows, and the
  columns contain the variance components for the subject level intercepts.
  Note that only the intercept variances (and not the co-variances) are
  returned.}
  \item{\code{gamma_int_subj}}{A list containing one matrix per subject
  denoting the subject level intercepts of the Multinomial logistic
  regression modeling the transition probabilities over the iterations of the
  hybrid Metropolis within Gibbs sampler. The iterations of the sampler are
  contained in the rows, and the columns contain the subject level
  intercepts.}
  \item{\code{gamma_naccept}}{A matrix containing the number of accepted
  draws at the subject level RW Metropolis step for each set of parameters of
  the transition probabilities. The subjects are contained in the rows, and
  the columns contain the sets of parameters.}
  \item{\code{emiss_mu}}{A list containing one matrix per dependent
  variable, denoting the (fixed over subjects) means of the Normal emission
  distribution of each dependent variable over the iterations of the Gibbs
  sampler. The iterations of the sampler are contained in the rows of the
  matrix, and the columns contain the group level emission means.}
  \item{\code{emiss_sd}}{A list containing one matrix per dependent
  variable, denoting the (fixed over subjects) standard deviation of the
  Normal emission distributions over the iterations of the Gibbs sampler. The
  iterations of the sampler are contained in the rows of the matrix, and the
  columns contain the group level emission variances.}
  \item{\code{input}}{Overview of used input specifications: the number of
  states \code{m}, the number of used dependent variables \code{n_dep}, the
  number of iterations \code{J} and the specified burn in period
  \code{burn_in} of the hybrid Metropolis within Gibbs sampler, the number of
  subjects \code{n_subj}, the observation length for each subject
  \code{n_vary}, and the column names of the dependent variables
  \code{dep_labels}.}
  \item{\code{input}}{Overview of used input specifications: the distribution
  type of the observations \code{data_distr}, the number of
  states \code{m}, the number of used dependent variables \code{n_dep}, the number of iterations
  \code{J} and the specified burn in period \code{burn_in} of the hybrid
  Metropolis within Gibbs sampler, the number of subjects \code{n_subj}, the
  observation length for each subject \code{n_vary}, and the column names of
  the dependent variables \code{dep_labels}.}
  \item{\code{sample_path}}{A list containing one matrix per subject with the
  sampled hidden state sequence over the hybrid Metropolis within Gibbs
  sampler. The time points of the dataset are contained in the rows, and the
  sampled paths over the iterations are contained in the columns. Only
  returned if \code{return_path = TRUE}. }
  }
}
\description{
\code{mHMM_f} fits a multilevel (also known as mixed or random effects)
hidden Markov model (HMM) to intense longitudinal data with (ONLY) continuous
(i.e., normally distributed), observations of multiple subjects using
Bayesian estimation, and creates an object of class \code{mHMM_f}. By using a
multilevel framework, we allow for heterogeneity in the transition model
parameters between subjects, while estimating one overall HMM. The function
includes the possibility to add covariates at level 2 (i.e., at the subject
level) and have varying observation lengths over subjects. For a short
description of the package see \link{mHMMbayes}. See
\code{vignette("tutorial-mhmm")} for an introduction to multilevel hidden
Markov models and the package, and see \code{vignette("estimation-mhmm")} for
an overview of the used estimation algorithms.
}
\details{
Covariates specified in \code{xx} can either be dichotomous or continuous
variables. Dichotomous variables have to be coded as 0/1 variables.
Categorical or factor variables can as yet not be used as predictor
covariates. The user can however break up the categorical variable in
multiple dummy variables (i.e., dichotomous variables), which can be used
simultaneously in the analysis. Continuous predictors are automatically
centered. That is, the mean value of the covariate is subtracted from all
values of the covariate such that the new mean equals zero. This is done such
that the presented probabilities in the output (i.e., for the population
transition probability matrix and population emission probabilities)
correspond to the predicted probabilities at the average value of the
covariate(s).
}
\examples{
###### Example on continuous simulated data
\donttest{ # simulating multivariate continuous data
n_t     <- 100
n       <- 10
m       <- 3
n_dep   <- 2

gamma   <- matrix(c(0.8, 0.1, 0.1,
                    0.2, 0.7, 0.1,
                    0.2, 0.2, 0.6), ncol = m, byrow = TRUE)

emiss_distr <- list(matrix(c( 50, 10,
                              100, 10,
                              150, 10), nrow = m, byrow = TRUE),
                    matrix(c(5, 2,
                             10, 5,
                             20, 3), nrow = m, byrow = TRUE))

data_cont <- sim_mHMM(n_t = n_t, n = n, data_distr = 'continuous',
                      gen = list(m = m, n_dep = n_dep),
                      gamma = gamma, emiss_distr = emiss_distr,
                      var_gamma = .1, var_emiss = c(5^2, 0.2^2))

# Specify hyper-prior for the continuous emission distribution
manual_prior_emiss <- prior_emiss_cont_f(
                        gen = list(m = m, n_dep = n_dep),
                        emiss_mu0 = list(matrix(c(30, 70, 170), nrow = 1),
                                         matrix(c(7, 8, 18), nrow = 1)),
                        emiss_K0 = list(1, 1),
                        emiss_V =  list(rep(5^2, m), rep(0.5^2, m)),
                        emiss_nu = list(1, 1))

# Run the model on the simulated data:
# Note that for reasons of running time, J is set at a ridiculous low value.
# One would typically use a number of iterations J of at least 1000,
# and a burn_in of 200.
out_3st_cont_sim <- mHMM_f(s_data = data_cont$obs,
                         data_distr = 'continuous',
                         gen = list(m = m, n_dep = n_dep),
                         start_val = c(list(gamma), emiss_distr),
                         emiss_hyp_prior = manual_prior_emiss,
                         mcmc = list(J = 11, burn_in = 5))

}


}
\references{
\insertRef{rabiner1989}{mHMMbayes}

\insertRef{scott2002}{mHMMbayes}

\insertRef{altman2007}{mHMMbayes}

\insertRef{rossi2012}{mHMMbayes}

\insertRef{zucchini2017}{mHMMbayes}
}
\seealso{
\code{\link{sim_mHMM}} for simulating multilevel hidden Markov data,
  \code{\link{vit_mHMM}} for obtaining the most likely hidden state sequence
  for each subject using the Viterbi algorithm, \code{\link{obtain_gamma}}
  and \code{\link{obtain_emiss}} for obtaining the transition or emission
  distribution probabilities of a fitted model at the group or subject level,
  and \code{\link{plot.mHMM}} for plotting the posterior densities of a
  fitted model.
}
