% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMM_mnl.R
\name{mHMM_mnl}
\alias{mHMM_mnl}
\title{Multilevel hidden markov model using Bayesian estimation}
\usage{
mHMM_mnl(s_data, gen, xx = NULL, start_val, gamma_sampler = NULL,
  emis_sampler = NULL, gamma_hyp_prior = NULL, emis_hyp_prior = NULL,
  mcmc, return_path = FALSE)
}
\arguments{
\item{s_data}{A matrix containing the observations to be modelled, with one
row per observation. In the \code{s_data} matrix, the first column
indicates subject id number. Hence, the id number is repeated over rows
equal to the number of observations for that subject. The number of
observations can vary over subjects. The subsequent columns contain the
dependent variable(s). The total number of rows are equal to the sum over
the number of observations of each subject, and the number of columns are
equal to the number of dependent variables (\code{n_dep}) + 1.}

\item{gen}{List containing the following elements:
\itemize{\item{\code{m}: numeric vector with length 1 denoting the number
of states} \item{\code{n_dep}: numeric vector with length 1 denoting the
number of dependent variables} \item{\code{q_emis}: numeric vector with
length \code{n_dep} denoting the number of obseved catagories for the
catagorical emission distribution of each dependent variable.}}}

\item{start_val}{Startvalues for gamma and emis used for first run of the
forward algorithm}

\item{gamma_sampler}{List containing start values for mle estimates of pooled
data for gamma, int_mle0, scalar and weight for the overall ll in the
fractional likelihood, w}

\item{emis_sampler}{List containing start values for mle estimates of pooled
data for emis, emis_int_mle0, emis_scalar and weight for the overall ll in the
fractional likelihood, emis_w}

\item{gamma_hyp_prior}{List containing \code{mu0}, \code{K0}, \code{nu} and \code{V}, see details below}

\item{emis_hyp_prior}{List containing \code{emis_mu0}, \code{emis_K0}, \code{emis_nu} and \code{emis_V}, see details below}

\item{mcmc}{List of MCMC argument, containing the following elements:
\code{J} number of iterations of the MCMC algorithm. \code{burn_in} Burn in
period for the MCMC algorithm}

\item{return_path}{A logical scalar. Should the sampled state sequence obtianed at
each iteration and for each subject be returned by the function
(\code{sample_path = TRUE}) or not (\code{sample_path = FALSE}). This is
quite a large object! Can be used for local decoding purposes.}
}
\value{
A list of
}
\description{
This function is for an multilevel model, can handle covariates at level 2
(unlimited number), uses a hybrid metropolis within gibs sampler, where the
metropolis hastings sampler is a random walk, and does the forward backward
algorithm for all subjects in a sequential manner. Can handle varying
observation length over subjects
}
\details{
Here are the details of the function
\subsection{Equations}{
  The following hyper prior is used on each row \eqn{i} of gamma (\eqn{\Gamma_i}) at the population level:
  \deqn{\Gamma_i ~ (idd) MNL(int_i)}
  \deqn{int_i ~ N(mu_int_bar, V_int)}
  \deqn{mu_int_bar ~ N(mu0, \frac{1}{K0} * V_int)}
  \deqn{V_int ~ IW(nu, V)}
  where \eqn{\alpha_i} denotes \code{int_i}, \eqn{\bar{\mu}_\alpha} denotes \code{mu_int_bar},  \code{} \code{} \code{} \code{} \code{}

  The following hyper prior is used for the conditional emission distribution of each state at the population level:
  \deqn{\theta_i ~ (idd) MNL(emis_int_i)}
  \deqn{emis_int ~ N(mu_emis_int_bar, V_emis_int)}
  \deqn{mu_emis_int_bar ~ N(emis_mu0, (1 / emis_K0) * V_emis_int)}
  \deqn{V_emis_int 					~ IW(emis_nu, emis_V)}

}

\subsection{Covariates}{
  Here is some info on using covariates
}
}
\examples{
# specifying general model properties
m <- 2
n_dep <- 4
q_emis <- c(3, 2, 3, 2)

# specifying starting values
start.EM <- list(matrix(c(0.9, 0.05, 0.05, 0.05, 0.05, 0.9), byrow = TRUE,
                        nrow = m, ncol = q_emis[1]), # vocalizing patient
                 matrix(c(0.9, 0.1, 0.9, 0.1), byrow = TRUE, nrow = m,
                        ncol = q_emis[2]), # looking patient
                 matrix(c(0.05, 0.05, 0.9, 0.9, 0.05, 0.05), byrow = TRUE,
                        nrow = m, ncol = q_emis[3]), # vocalizing therapist
                 matrix(c(0.9, 0.1, 0.9, 0.1), byrow = TRUE, nrow = m,
                        ncol = q_emis[4])) # looking therapist
start.TM <- diag(.8, m)
start.TM[lower.tri(start.TM) | upper.tri(start.TM)] <- .2

# run a model without covariates
out <- mHMM_mnl(s_data = nonverbal, gen = list(m = m, n_dep = n_dep,
                q_emis = q_emis), start_val = c(as.vector(t(start.EM[[1]])),
                as.vector(t(start.EM[[2]])), as.vector(t(start.EM[[3]])),
                as.vector(t(start.EM[[4]])), as.vector(t(start.TM))),
                mcmc = list(J = 11, burn_in = 5))

# including covariates. Only the emission distribution for each of the 4
# dependent variables is predicted using standardized CDI change.
n_subj <- 10
xx <- rep(list(matrix(1, ncol = 1, nrow = n_subj)), (n_dep + 1))
for(i in 2:(n_dep + 1)){
  xx[[i]] <- cbind(xx[[i]], nonverbal_cov$std_CDI_change)
}
out2 <- mHMM_mnl(s_data = nonverbal, xx = xx, gen = list(m = m, n_dep = n_dep,
                q_emis = q_emis), start_val = c(as.vector(t(start.EM[[1]])),
                as.vector(t(start.EM[[2]])), as.vector(t(start.EM[[3]])),
                as.vector(t(start.EM[[4]])), as.vector(t(start.TM))),
                mcmc = list(J = 11, burn_in = 5))


}
