% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_emiss.R
\name{obtain_emiss}
\alias{obtain_emiss}
\title{Obtain the emission distribution proabilities for a fitted multilevel HMM}
\usage{
obtain_emiss(object, level = "group", burn_in = NULL)
}
\arguments{
\item{object}{An object of class \code{mHMM}, generated by the function
\code{\link{mHMM}}.}

\item{level}{String specifying if the returned emission distribution
proabilities should be at the group level (\code{level = "group"}), i.e.,
representing the average emission distribution proabilities over all
subjects, or at the subject level (\code{level = "subject"}). When the
latter is specified, the returned emission distribution proabilities are
given in a list with the number of elements equal to the number of subjects
analyzed.}

\item{burn_in}{An integer wich specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified (\code{burn_in = NULL}), the burn in period specified when
creating the \code{mHMM} object with the function \code{\link{mHMM}}
will be used.}
}
\value{
\code{obtain_emiss} returns the object \code{est_emiss}. Depending on
  the specification at the input variable \code{level}, \code{est_emiss} is
  either a list of matrices with the emission distribution proabilities at
  the group level (if \code{level = "group"}) for each dependent variable, or
  a list of lists, where for each dependent variable a list is returned with
  the number of elements equal to the number of subjects analyzed, if
  \code{level = 'subject'}). In the latter scenario, each matrix in the lower
  level list represents the subject specific emission distribution
  proabilities for a specific dependent variable.

#' @seealso \code{\link{mHMM}} for fitting the multilevel hidden Markov
model, creating the object \code{mHMM}.
}
\description{
\code{obtain_emiss} obtains the emission distribution proabilities (also
known as conditional probabilities) for a fitted multilevel hidden markov
model, for either the group level, i.e., representing the average emission
distribution proabilities over all subjects, or at the subject level,
returning the emission distribution proabilities for each subject.
}
