% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_emiss.R
\name{obtain_emiss_f}
\alias{obtain_emiss_f}
\title{Obtain the emission distribution probabilities for a fitted multilevel HMM}
\usage{
obtain_emiss_f(object, burn_in = NULL)
}
\arguments{
\item{object}{An object of class \code{mHMM_f}, generated by the function
\code{\link{mHMM_f}}.}

\item{burn_in}{An integer which specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified (\code{burn_in = NULL}), the burn in period specified when
creating the \code{mHMM_f} object will be used.}
}
\value{
\code{obtain_emiss_f} creates an object of the class
  \code{mHMM_emiss_f}. The output is a list of matrices with the fixed over
  subjects emission distribution for each dependent variable.
}
\description{
\code{obtain_emiss_f} obtains the emission distribution for an object
containing a fitted multilevel hidden Markov model with a fixed over subjects emission distribution.
}
\examples{
###### Currently no example
}
\seealso{
\code{\link{mHMM_f}} for fitting the
  multilevel hidden Markov model with a fixed emission distribution.
}
