% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_gamma.R
\name{obtain_gamma}
\alias{obtain_gamma}
\title{Obtain the transition proabilities gamma for a fitted multilevel HMM}
\usage{
obtain_gamma(object, level = "group", burn_in = NULL)
}
\arguments{
\item{object}{An object of class \code{mHMM}, generated by the function
\code{\link{mHMM_mnl}}.}

\item{level}{String specifying if the returned transition probability matrix
gamma should be at the group level (\code{level = "group"}), i.e.,
representing the average transition probability matrix over all subjects,
or at the subject level (\code{level = "subject"}). When the latter is
specified, the returned gamma is a list with the number of elements equal
to the number of subjects analyzed.}

\item{burn_in}{An integer wich specifies the number of iterations to discard
when obtaining the model parameter summary statistics. When left
unspecified (\code{burn_in = NULL}), the burn in period specified when
creating the \code{mHMM} object with the function \code{\link{mHMM_mnl}}
will be used.}
}
\value{
\code{obtain_gamma} returns the object \code{est_gamma} of the class
  \code{mHMM_gamma}. This object can be directly plotted using the function
  \code{plot.mHMM_gamma()}, or simply \code{plot()}. Depending on the
  specification at the input variable \code{level}, \code{est_gamma} is
  either a matrix with the transiton probabilities at the group level (if
  \code{level = "group"}), or a list of matrices (with the number of elements
  equal to the number of subjects analyzed, if \code{level = 'subject'}),
  where each matrix in the list represents a subject specific transition
  probability matrix.

#' @seealso \code{\link{mHMM_mnl}} for fitting the multilevel hidden Markov
model, creating the object \code{mHMM}.
}
\description{
\code{obtain_gamma} obtains the tranisitoin probability matrix for a
fitted multilevel hidden markov model, for either the group level, i.e.,
  representing the average transition probability matrix over all subjects,
  or at the subject level, returning the transition probability matrices for
  each subject.
}
