% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mHMM_emiss.R
\name{plot.mHMM_emiss}
\alias{plot.mHMM_emiss}
\title{Plotting the emission distribution for a fitted multilevel HMM}
\usage{
\method{plot}{mHMM_emiss}(
  x,
  subj_nr = NULL,
  by_state = TRUE,
  cat_lab,
  dep_lab,
  col,
  legend_cex = 1,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mHMM_emiss}, generated by the function
\code{\link{obtain_emiss}}.}

\item{subj_nr}{An integer specifying for which specific subject the
transition probability matrix should be plotted. Only required if the input
object represents the subject specific transition probability matrices.}

\item{by_state}{An optional logic variable. It indicates how user would like
to display the plots. It is set to \code{TRUE} by default and it
display plots state wise. If dependent variable wise presentation
is preferred \code{by_state} should be set to \code{FALSE}
(\code{by_state=FALSE})}

\item{cat_lab}{Optional list containing \code{n_dep} elements used when
plotting the posterior densities of the emission probabilities. Each
element in the list is a vector of strings denoting the labels of the
categorical outcome values for each dependent variable. Automatically
generated when not provided.}

\item{dep_lab}{Optional vector of strings when plotting the posterior
densities of the emission probabilities with length \code{n_dep}, denoting
the labels of the dependent variables. Automatically obtained from the
input object \code{x} when not specified.}

\item{col}{An optional vector of the size of the number of categories of the
dependent variable that has them the most,specifying the used colors in the
\code{\link[graphics]{barplot}}.}

\item{legend_cex}{A numerical value giving the amount by which
to the default.}

\item{...}{Arguments to be passed to methods
(see \code{\link[graphics]{barplot}})}
}
\value{
\code{plot.mHMM_gamma} returns a plot of the emission probability
  matrix. Depending on whether the input object represents list of emission
  probability matrices at the group level or the subject specific list of
  emission probability matrices, the returned plot represents either
  the group emission probability distribution for each state, or the emission
  probability distribution for each state of given subject,
  specified by \code{subject_nr}.
}
\description{
\code{plot.mHMM_emiss} plots the emission probability distribution for
a fitted multilevel hidden Markov model.
The plotted emission probability distributions either represents
the probabilities at the group level, i.e., representing the average emission
probability distribution over all subjects, or at the subject level.
In case of the latter, the user has to specify for which subject the emission
probability matrix should be plotted.
}
\examples{
\donttest{
 #Discrete case
m <- 3 #number of states
n_dep <- 4 #number of dependent variables used to infer the hidden states
q_emiss <- c(3, 2, 3, 2) #number of categorical outcomes of
                          each of dependent variables
start_TM <- diag(.8, m) #transition matrix
start_TM
start_TM[lower.tri(start_TM) | upper.tri(start_TM)] <- .2
start_EM <- list(matrix(c(0.05, 0.90, 0.05,
                          0.90, 0.05, 0.05,
                         0.60, 0.30, 0.10), byrow = TRUE,
                       nrow = m, ncol = q_emiss[1]), # vocalizing patient
                matrix(c(0.1, 0.9,
                         0.1, 0.9,
                         0.5, 0.5), byrow = TRUE, nrow = m,
                       ncol = q_emiss[2]), # looking patient
                matrix(c(0.90, 0.05, 0.05,
                         0.05, 0.90, 0.05,
                         0.60, 0.30, 0.10), byrow = TRUE,
                       nrow = m, ncol = q_emiss[3]), # vocalizing therapist
                matrix(c(0.1, 0.9,
                         0.1, 0.9,
                         0.5, 0.5), byrow = TRUE, nrow = m,
                       ncol = q_emiss[4])) # looking therapist

out_3st <- mHMM(s_data = nonverbal,
                gen = list(m = m, n_dep = n_dep, q_emiss = q_emiss),
                start_val = c(list(start_TM), start_EM),
                mcmc = list(J = 10, burn_in = 5))

# creating object of obtain_emiss class
group_emiss3<-obtain_emiss(out_3st,level = "group")

subject_emiss3<-obtain_emiss(out_3st,level = "subject")

#plotting
plot(x=subject_emiss3,subj_nr = 2)

plot(x=group_emiss3)

plot(x=group_emiss3,by_state=FALSE)

#continuous case
n_t <- 100
n <- 10
m <- 3
J = 11
burn_in = 5
n_dep2 <- 2

emiss_distr <- list(matrix(c( 5, 1,
10, 1,
15, 1), nrow = m, byrow = TRUE),
matrix(c(0.5, 0.1,
        1.0, 0.2,
        2.0, 0.1), nrow = m, byrow = TRUE))
gamma <- matrix(c(0.8, 0.1, 0.1,
0.2, 0.6, 0.2,
0.1, 0.2, 0.7), ncol = m, byrow = TRUE)

#simulate data with continuous distribution for 2 dependent variables

data_cont <- sim_mHMM(n_t = n_t, n = n, m = m, n_dep = n_dep2,
 data_distr = c('continuous','continuous'),
gamma = gamma, emiss_distr = emiss_distr, var_gamma = .1,
var_emiss = c(.5, 0.01))
# Specify hyper-prior for the continuous emission distribution
hyp_pr <- list(
 emiss_mu0 = list(matrix(c(3,7,17), nrow = 1), matrix(c(0.7, 0.8, 1.8),
  nrow = 1)),
 emiss_K0  = list(1, 1),
 emiss_nu  = list(1, 1),
 emiss_V   = list(rep(2, m), rep(1, m)),
 emiss_a0  = list(rep(1, m), rep(1, m)),
 emiss_b0  = list(rep(1, m), rep(1, m)))
# Fit the mHMM_cont on 2 dep variable data:
out_3st_cont_sim <- mHMM_cont(s_data = data_cont$obs,
                             gen = list(m = m, n_dep = n_dep2),
                             start_val = c(list(gamma), emiss_distr),
                             emiss_hyp_prior = hyp_pr,
                             mcmc = list(J = J, burn_in =burn_in,
                             show_progress = FALSE))

emiss1_g_cont <- obtain_emiss(out_3st_cont_sim)
plot.mHMM_emiss(emiss1_g_cont)

}

}
\seealso{
\code{\link{mHMM}} for fitting the multilevel hidden Markov
  model, creating the object \code{mHMM}, and \code{\link{obtain_emiss}} to
  obtain the transition probabilities gamma for a fitted multilevel HMM,
  creating the object \code{mHMM_gamma}.
}
