% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mHMM_emiss.R
\name{plot.mHMM_emiss}
\alias{plot.mHMM_emiss}
\title{Plotting the emission distribution for a fitted multilevel HMM}
\usage{
\method{plot}{mHMM_emiss}(x, subj_nr = NULL, by_state = TRUE, group_labs, cat_labs, col)
}
\arguments{
\item{x}{An object of class \code{mHMM_emiss}, generated by the function
\code{\link{obtain_emiss}}.}

\item{subj_nr}{An integer specifying for which specific subject the
transition probability matrix should be plotted. Only required if the input
object represents the subject specific transition probability matrices.}

\item{by_state}{An optional logic variable. It indicates how user would like
to display the plots. It is set to \code{TRUE} by default and it
display plots state wise. If dependent variable wise presentation
is prefered \code{by_state} should be set to \code{FALSE}
(\code{by_state=FALSE})}

\item{group_labs}{An optional vector of length \code{n_dep} with names
for each dependent variable}

\item{col}{An optional vector of length \code{max(q_emiss)} (i.e., where
\code{max(q_emiss)} denotes the number of categories of dependent variable
 with the most categories) specifying the used colors for the plot.}

\item{q_emiss}{A vector of length \code{n_dep} representing numbers of
categorical outcomes for each of dependent variables.}
}
\value{
\code{plot.mHMM_gamma} returns a plot of the emission probability
  matrix. Depending on whether the input object represents list of emission
  probability matrices at the group level or the subject specific list of
  emission probability matrices, the returned plot represents either
  the group emission probability distribution for each state, or the emission
  probability distribution for each state of given subject,
  specified by \code{subject_nr}.
}
\description{
\code{plot.mHMM_emiss} plots the emission probability distribution for
a fitted multilevel hidden Markov model.
The plotted emission probability distributions either represents
the probabilities at the group level, i.e., representing the average emission
probability distribution over all subjects, or at the subject level.
In case of the latter, the user has to specify for which subject the emission
probability matrix should be plotted.
}
\examples{
\donttest{
m <- 3 #number of states
n_dep <- 4 #number of dependent variables used to infer the hidden states
q_emiss <- c(3, 2, 3, 2) #number of categorical outcomes of
                          each of dependent variables
start_TM <- diag(.8, m) #transition matrix
start_TM
start_TM[lower.tri(start_TM) | upper.tri(start_TM)] <- .2
start_EM <- list(matrix(c(0.05, 0.90, 0.05,
                          0.90, 0.05, 0.05,
                         0.60, 0.30, 0.10), byrow = TRUE,
                       nrow = m, ncol = q_emiss[1]), # vocalizing patient
                matrix(c(0.1, 0.9,
                         0.1, 0.9,
                         0.5, 0.5), byrow = TRUE, nrow = m,
                       ncol = q_emiss[2]), # looking patient
                matrix(c(0.90, 0.05, 0.05,
                         0.05, 0.90, 0.05,
                         0.60, 0.30, 0.10), byrow = TRUE,
                       nrow = m, ncol = q_emiss[3]), # vocalizing therapist
                matrix(c(0.1, 0.9,
                         0.1, 0.9,
                         0.5, 0.5), byrow = TRUE, nrow = m,
                       ncol = q_emiss[4])) # looking therapist

out_3st <- mHMM(s_data = nonverbal,
                gen = list(m = m, n_dep = n_dep, q_emiss = q_emiss),
                start_val = c(list(start_TM), start_EM),
                mcmc = list(J = 10, burn_in = 5))

# creating object of obtain_emiss class
group_emiss3<-obtain_emiss(out_3st,level = "group")

subject_emiss3<-obtain_emiss(out_3st,level = "subject")

# setting new dependent variables' labels
group_labs<-c("voc_pat","look_pat","voc_ther","look_ther")

#plotting
plot(x=subject_emiss3, group_labs = group_labs,subj_nr = 2)

plot(x=group_emiss3)

plot(x=group_emiss3,group_labs=group_labs,by_state=FALSE)


}

}
\seealso{
\code{\link{mHMM}} for fitting the multilevel hidden Markov
  model, creating the object \code{mHMM}, and \code{\link{obtain_emiss}} to
  obtain the transition probabilities gamma for a fitted multilevel HMM,
  creating the object \code{mHMM_gamma}.
}
