% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mHMM_gamma.R
\name{plot.mHMM_gamma}
\alias{plot.mHMM_gamma}
\title{Plotting the transition proabilities gamma for a fitted multilevel HMM}
\usage{
\method{plot}{mHMM_gamma}(x, subj_nr = NULL, cex = 0.8, col, hide, ...)
}
\arguments{
\item{x}{An object of class \code{mHMM_gamma}, generated by the function
\code{\link{obtain_gamma}}.}

\item{subj_nr}{An integer specifying for which specific subject the
transition probability matrix should be plotted. Only required if the input
object represents the subject specific transition probability matrices.}

\item{cex}{numeric, scaling of fonts of category labels. When not specified,
defaults to \code{cex = 0.8}.}

\item{col}{vector of colors of the stripes.}

\item{hide}{logical, should particlular stripe be plotted. When not
specified, omits the lines representing a value of exactly zero.}

\item{...}{Arguments to be passed to alluvial (see \code{\link[alluvial]{alluvial}})}
}
\value{
\code{plot.mHMM_gamma} returns a plot of the transition probability
  matrix. Depending on whether the input object represents the transiton
  probabilities at the group level or the subject specific transition
  probability matrices, the returned plot represents either the group
  transition probality matrix, or the transition probability matrix for a
  given subject, specified by \code{subject_nr}.
}
\description{
\code{plot.mHMM_gamma} plots the tranisitoin probability matrix for a fitted
multilevel hidden markov model, by means of an alluvial plot (also known as
sankey diagram or riverplot) using the R package \code{alluvial}. The plotted
transition probability matrix either represents the probabilities at the
group level, i.e., representing the average transition probability matrix
over all subjects, or at the subject level. In case of the latter, the user
has to specify for which subject the transition probability matrix should be
plotted.
}
\seealso{
\code{\link{mHMM}} for fitting the multilevel hidden Markov
  model, creating the object \code{mHMM}, and \code{\link{obtain_gamma}} to
  obtain the the transition proabilities gamma for a fitted multilevel HMM,
  creating the object \code{mHMM_gamma}.
}
