% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mHMM.R
\name{sim_mHMM}
\alias{sim_mHMM}
\title{Simulate data using a multilevel hidden markov model}
\usage{
sim_mHMM(T, n, m, pr, gamma, emiss_distr, var_gamma = 1, var_emiss = 1,
  return_ind_par = FALSE)
}
\arguments{
\item{T}{The length of the observed sequence to be simulated for each
subject. To only simulate subject specific transition probability matrices
gamma and emission distributions (and no data), set \code{T} to 0.}

\item{n}{The number of subjects for which data is simulated.}

\item{m}{The number of hidden states in the HMM used for
simulating data.}

\item{pr}{The number of categories of the simulated observations.}

\item{gamma}{A matrix with \code{m} rows and \code{m} columns containing the
average population transition probability matrix used for simulating the
data. That is, the probability to switch from hidden state \emph{i} (row
\emph{i}) to hidden state \emph{j} (column  \emph{j}).}

\item{emiss_distr}{A matrix with \code{m} rows and \code{pr} columns
containing the average population emission distribution of the
(categorical) observations given the hidden states. That is, the
probability of observing category \emph{k} (column \emph{k}) in state
\emph{i} (row \emph{i}).}

\item{var_gamma}{An integer denoting the variance between subjects in the
transition probability matrix. Note that this value corresponds to the
variance of the parameters of the multinomial distribution (i.e., the
intercepts of the regression equation of the multinomial distribution used
to sample the transition probability matrix), see details below. In
addition, only one variance value can be specified for the complete
transition probability matrix, hence the variance is assumed fixed across
all components. The default equals 1, which corresponds to quite some
variation between subjects. A less extreme value would be 0.5. If one wants
to simulate data from exactly the same HMM for all subjects, var_gamma
should be set to 0.}

\item{var_emiss}{An integer denoting the variance between subjects in the
emission distribution. Note that this value corresponds to the variance of
the parameters of the multinomial distribution (i.e., the intercepts of the
regression equation of the multinomial distribution used to sample the
components of the emission distribution), see details below.  In addition,
only one variance value can be specified for the complete emission
distribution, hence the variance is assumed fixed across all components.
The default equals 1, which corresponds to quite some variation between
subjects. A less extreme value would be 0.5. If one wants to simulate data
from exactly the same HMM for all subjects, var_emmis should be set to
0.}

\item{return_ind_par}{A logical scalar. Should the subject specific
  transition probability matrix gamma and emission probability matrix be
  returned by the function (\code{return_ind_par = TRUE}) or not
  (\code{return_ind_par = FALSE}). The default equals \code{return_ind_par =
  FALSE}.

Note: the subject specific) initial state distributions (i.e., the
probability of each of the states at the first time point) needed to simulate
the data are obtained from the stationary distributions of the subject
specific transition probability matrices gamma.}
}
\description{
This function simulates data for multiple subjects, for which the data have
categorical observations that follow a hidden Markov model (HMM) with an
multilevel structure. The multilevel structure means that each subject can
have it's own set of parameters, and that the parameters at the subject
level (level 1) are tied together by a population distribution at level 2 for
each of the corresponding parameters. The shape of the population distribution
for each of the parameters is a normal (i.e., Gaussian) distribution. In
simulating the data, having a multilevel structure means that the parameters
for each subject are sampled from the population level distribution of the
corresponding parameter. The user specifies the population distribution for
each parameter: the average population transition probability matrix and its
variance, and the average population emission distribution and its variance.
For now, the variance is assumed fixed for all components of the transition
probability matrix and for all components of the emissiondistribution, and
the simulated data can only consist of one dependent variable.
}
\examples{
# simulating data for 10 subjects with each 100 observations
T <- 100
n <- 10
m <- 3
pr <- 4
gamma <- matrix(c(0.8, 0.1, 0.1,
                  0.2, 0.7, 0.1,
                  0.2, 0.2, 0.6), ncol = m, byrow = TRUE)
emiss_distr <- matrix(c(0.5, 0.5, 0.0, 0.0,
                        0.1, 0.1, 0.8, 0.0,
                        0.0, 0.0, 0.1, 0.9), nrow = m, ncol = pr, byrow = TRUE)
data1 <- sim_mHMM(T = T, n = n, m = m, pr = pr, gamma = gamma, emiss_distr = emiss_distr,
                  var_gamma = 1, var_emiss = 1)


# simulating subject specific transition probability matrices and emission distributions only
T <- 0
n <- 5
m <- 3
pr <- 4
gamma <- matrix(c(0.8, 0.1, 0.1,
                  0.2, 0.7, 0.1,
                  0.2, 0.2, 0.6), ncol = m, byrow = TRUE)
emiss_distr <- matrix(c(0.5, 0.5, 0.0, 0.0,
                        0.1, 0.1, 0.8, 0.0,
                        0.0, 0.0, 0.1, 0.9), nrow = m, ncol = pr, byrow = TRUE)
data2 <- sim_mHMM(T = T, n = n, m = m, pr = pr, gamma = gamma, emiss_distr = emiss_distr,
                  var_gamma = 1, var_emiss = 1)
data2

data3 <- sim_mHMM(T = T, n = n, m = m, pr = pr, gamma = gamma, emiss_distr = emiss_distr,
                  var_gamma = .5, var_emiss = .5)
data3
}
