% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mHMM.R
\name{summary.mHMM}
\alias{summary.mHMM}
\title{Summary of a \code{mHMM} object}
\usage{
\method{summary}{mHMM}(object, ...)
}
\arguments{
\item{object}{A \code{mHMM} object.}

\item{...}{Other parameters passed down to \code{summary()}.}
}
\description{
Summary of a \code{mHMM} object
}
